% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsources.210324.R
\name{data_summary}
\alias{data_summary}
\title{Summarize Data by Group}
\usage{
data_summary(data, varname, groupnames)
}
\arguments{
\item{data}{A data frame containing the data to be summarized.}

\item{varname}{The name of the variable to summarize.}

\item{groupnames}{A vector of names of grouping variables.}
}
\value{
A data frame with the mean and standard deviation of the specified variable for each group.
}
\description{
This function provides a summary (mean and standard deviation) of a specified variable within a data frame, grouped by one or more factors.
}
\examples{
data <- data.frame(group = rep(c("A", "B"), each = 5), value = rnorm(10))
data_summary(data, varname = "value", groupnames = "group")

}
