% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsources.210324.R
\name{getTaxonomy}
\alias{getTaxonomy}
\title{Retrieve Taxonomy Information for Given OTUs}
\usage{
getTaxonomy(otus, tax_tab, level, taxRanks, na_str = c("NA"))
}
\arguments{
\item{otus}{Vector of OTUs for which taxonomy information is required.}

\item{tax_tab}{Taxonomy table containing taxonomic information for each OTU.}

\item{level}{Taxonomic level at which information is sought.}

\item{taxRanks}{Vector of all possible taxonomic ranks.}

\item{na_str}{Character string to represent missing data in taxonomy.}
}
\value{
A data frame with taxonomy information for the specified OTUs.
}
\description{
Retrieve Taxonomy Information for Given OTUs
}
\examples{
# Example usage pending specific data structure

}
