% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsources.210324.R
\name{getTaxonomyAll}
\alias{getTaxonomyAll}
\title{Retrieve Comprehensive Taxonomy Information}
\usage{
getTaxonomyAll(otus, tax_tab, level, na_str = c("NA"))
}
\arguments{
\item{otus}{Vector of OTUs for taxonomy information.}

\item{tax_tab}{Taxonomy table with taxonomic details.}

\item{level}{Desired taxonomic level.}

\item{na_str}{String to represent missing data.}
}
\value{
A data frame with comprehensive taxonomy information.
}
\description{
Retrieve Comprehensive Taxonomy Information
}
\examples{
# Example usage pending specific data structure

}
