% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_Aldex2_markdown.R
\name{Go_Aldex2_markdown}
\alias{Go_Aldex2_markdown}
\title{Differential Abundance Analysis Using ALDEx2}
\usage{
Go_Aldex2_markdown(
  psIN,
  cate.outs = NULL,
  cont.outs = NULL,
  cate.conf = NULL,
  cont.conf = NULL,
  orders = NULL
)
}
\arguments{
\item{psIN}{The phyloseq object containing OTU/ASV counts and associated sample data.}

\item{cate.outs}{A vector of column names in the sample data of 'psIN' representing categorical outcome variables.}

\item{cont.outs}{A vector of column names in the sample data of 'psIN' representing continuous outcome variables.}

\item{cate.conf}{A vector of column names in the sample data of 'psIN' representing categorical confounding variables.}

\item{cont.conf}{A vector of column names in the sample data of 'psIN' representing continuous confounding variables.}

\item{orders}{Optional ordering for the levels of the categorical variables.}
}
\value{
A list of data frames, each representing the results of the differential abundance analysis for one outcome variable.
Each data frame includes statistics like effect size, p-values, and adjusted p-values, along with taxonomic information for each feature.
}
\description{
This function performs differential abundance analysis using the ALDEx2 tool on a given phyloseq object.
It allows for the analysis of both categorical and continuous outcomes, with optional inclusion of confounding factors.
}
\examples{
# Assuming 'ps' is a phyloseq object with appropriate data
results <- Go_Aldex2_markdown(ps,
                             cate.outs = c("Group"),
                             cont.outs = c("Age"),
                             cate.conf = c("Gender"),
                             cont.conf = NULL,
                             orders = c("Control", "Treatment"))

}
