% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_ancom_plot.R
\name{Go_ancom_plot}
\alias{Go_ancom_plot}
\title{Generate ANCOM Log Fold Change Plots}
\usage{
Go_ancom_plot(
  project,
  file_path,
  files,
  type = "taxonomy",
  mycols = NULL,
  name,
  overlaps = 10,
  font,
  height,
  width
)
}
\arguments{
\item{project}{A string representing the name of the project.}

\item{file_path}{A string indicating the path to the input files.}

\item{files}{A pattern string to identify the specific CSV files to be processed.}

\item{type}{The type of analysis data, default is "taxonomy".}

\item{mycols}{Optional, a vector of colors for plot aesthetics.}

\item{name}{An optional string to define a specific naming convention for output files.}

\item{overlaps}{Integer, the maximum number of labels to allow overlapping.}

\item{font}{The font size for plot text.}

\item{height}{The height of the output plot.}

\item{width}{The width of the output plot.}
}
\value{
Generates PDF files containing ANCOM log fold change plots for each metadata variable comparison.
}
\description{
This function creates bar plots showing log fold changes from ANCOM results for each metadata variable comparison.
}
\details{
The function reads in CSV files containing ANCOM results, processes the data to highlight significant differences, and then generates bar plots showing the log fold changes for the comparisons of interest. These plots are saved as PDF files.
}
\examples{
# Example usage:
Go_ancom_plot(project = "MyProject", file_path = "data/", files = "ancom_results*.csv",
              type = "taxonomy", mycols = c("red", "blue"), name = "experiment1",
              overlaps = 10, font = 12, height = 6, width = 8)

}
