% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_barchart_v35.R
\name{Go_barchart}
\alias{Go_barchart}
\title{Create Bar Charts from Phyloseq Data}
\usage{
Go_barchart(
  psIN,
  cate.vars,
  project,
  taxanames,
  orders = NULL,
  simple = FALSE,
  mycols = NULL,
  relative = T,
  y_axis = NULL,
  x_label = NULL,
  facet = NULL,
  legend = "bottom",
  cutoff = 0.005,
  name = NULL,
  ncol = NULL,
  mark = NULL,
  height,
  width
)
}
\arguments{
\item{psIN}{Phyloseq object containing the data for analysis.}

\item{cate.vars}{Categorical variables for plotting.}

\item{project}{Name of the project or analysis.}

\item{taxanames}{Taxonomic levels for aggregation and analysis.}

\item{orders}{Custom order of factors in the plots, if applicable.}

\item{simple}{Logical value to choose between simple and detailed plotting.}

\item{mycols}{Color palette for the bars in the chart.}

\item{relative}{Logical value to select between relative (TRUE) or absolute (FALSE) abundance.}

\item{y_axis}{Label for the Y-axis.}

\item{x_label}{Label for the X-axis.}

\item{facet}{Faceting variable for the plot.}

\item{legend}{Position of the legend ("bottom" or "right").}

\item{cutoff}{Threshold for filtering low-abundance taxa.}

\item{name}{Optional name for the analysis.}

\item{ncol}{Number of columns for facet wrapping in the plot.}

\item{height}{Height of the output plot.}

\item{width}{Width of the output plot.}
}
\value{
Saves bar chart plots as PDF files in a specified directory and outputs tables
containing taxa abundance data. The function supports various customization options for the plots.
}
\description{
This function generates bar charts from a Phyloseq object, allowing for analysis
of taxonomic abundance at various levels. It supports both relative and absolute
abundance measures and offers customization options like color, faceting, and ordering.
}
\details{
The function preprocesses the Phyloseq object, potentially aggregating data at a specified taxonomic level,
and then generates bar charts for each taxonomic level specified. It handles relative and absolute abundance
data and allows the user to specify various aesthetic and layout options for the charts.
}
\examples{
# psIN is a Phyloseq object
# Example usage:
Go_barchart(psIN = psIN,
            cate.vars = c("Treatment", "Condition"),
            project = "MyProject",
            taxanames = c("Phylum", "Genus"),
            orders = c("Control", "Treatment"),
            relative = TRUE,
            y_axis = "Abundance",
            x_label = "SampleID",
            facet = "Treatment",
            legend = "bottom",
            cutoff = 0.005,
            name = "Analysis1",
            ncol = 2,
            height = 7,
            width = 10)

}
