% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_barchart_markdown.R
\name{Go_barchart_markdown}
\alias{Go_barchart_markdown}
\title{Generate Bar Charts for Microbiome Data}
\usage{
Go_barchart_markdown(
  psIN,
  cate.vars,
  project,
  taxanames,
  orders = NULL,
  simple = FALSE,
  mycols = NULL,
  relative = T,
  y_axis = NULL,
  x_label = NULL,
  facet = NULL,
  legend = "bottom",
  cutoff = 0.005,
  name = NULL,
  ncol = NULL
)
}
\arguments{
\item{psIN}{Phyloseq object containing the OTU/ASV counts and associated sample data.}

\item{cate.vars}{Vector of column names in the sample data of 'psIN' representing categorical variables for analysis.}

\item{project}{Name or identifier for the project or dataset being analyzed.}

\item{taxanames}{Vector of taxonomic ranks to be included in the analysis.}

\item{orders}{Optional ordering for the levels of the categorical variables.}

\item{simple}{Boolean indicating if a simplified version of the bar chart should be generated.}

\item{mycols}{Color palette for the bar chart.}

\item{relative}{Boolean indicating whether relative abundance should be used (default TRUE).}

\item{y_axis}{Label for the y-axis.}

\item{x_label}{Label for the x-axis, default is "SampleIDfactor".}

\item{facet}{Optional parameter for facetting the plot.}

\item{legend}{Position of the legend in the plot.}

\item{cutoff}{Cutoff value for filtering low-abundance taxa.}

\item{name}{Optional name for the analysis output.}

\item{ncol}{Number of columns for facetting.}
}
\value{
An invisible ggplot object representing the bar chart.
}
\description{
This function creates bar charts for microbiome data analysis from a Phyloseq object. It supports various options including taxonomic aggregation, relative abundance, customization of colors, and facetting.
}
\examples{
# Assuming 'ps' is a phyloseq object with appropriate data
Go_barchart_markdown(ps, cate.vars = c("Group"), project = "my_project",
                     taxanames = c("Phylum", "Genus"), relative = TRUE)

}
