% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_bdiv_V34.R
\name{Go_bdiv}
\alias{Go_bdiv}
\title{Generate Biodiversity Plots for Microbiome Data}
\usage{
Go_bdiv(
  psIN,
  cate.vars,
  project,
  orders,
  distance_metrics,
  cate.conf = NULL,
  plot = "PCoA",
  ellipse = TRUE,
  statistics = TRUE,
  mycols = NULL,
  paired = NULL,
  combination = NULL,
  shapes = NULL,
  ID = NULL,
  facet = NULL,
  name = NULL,
  addnumber = TRUE,
  height,
  width
)
}
\arguments{
\item{psIN}{Phyloseq object containing microbiome data.}

\item{cate.vars}{Categorical variables to be used for grouping in the plot.}

\item{project}{Name of the project for labeling.}

\item{orders}{Ordering of factors in the plot.}

\item{distance_metrics}{Vector of distance metrics to be used.}

\item{cate.conf}{Additional confounder variables, if any.}

\item{plot}{Type of plot to generate, e.g., "PCoA".}

\item{ellipse}{Logical indicating whether to include ellipses.}

\item{statistics}{Logical indicating whether to include statistical tests.}

\item{mycols}{Custom color palette.}

\item{paired}{Variable for paired analysis, if applicable.}

\item{combination}{Number of groups to combine for comparison.}

\item{shapes}{Shape of points in the plot.}

\item{ID}{Identifier for labeling points.}

\item{facet}{Faceting variable for the plot.}

\item{name}{Optional name for the output plot.}

\item{addnumber}{Logical indicating whether to add sample numbers to groups.}

\item{height}{Height of the plot.}

\item{width}{Width of the plot.}
}
\value{
A PDF file containing the biodiversity plot(s).
}
\description{
Generate Biodiversity Plots for Microbiome Data
}
\details{
This function creates biodiversity plots such as PCoA for microbiome data, facilitating comparisons across different conditions or factors. The function supports various distance metrics and provides options for customization.
}
\examples{
Go_bdiv(psIN = ps_object,
        cate.vars = c("Condition", "Treatment"),
        project = "MyMicrobiomeStudy",
        orders = c("Condition1", "Condition2"),
        distance_metrics = c("bray", "unifrac"),
        cate.conf = "AgeGroup",
        plot = "PCoA",
        ellipse = TRUE, or group names
        statistics = TRUE,
        mycols = c("blue", "red"),
        paired = "PatientID",
        combination = 2,
        shapes = 16,
        ID = "SampleID",
        facet = "Group",
        name = "BiodivPlot",
        addnumber = TRUE,
        height = 6,
        width = 8)

}
