% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_bdiv_V36.R
\name{Go_bdivPM}
\alias{Go_bdivPM}
\title{Generate Biodiversity Plots for Microbiome Data}
\usage{
Go_bdivPM(
  psIN,
  cate.vars,
  project,
  orders,
  distance_metrics,
  cate.conf = NULL,
  plot = "PCoA",
  ellipse = TRUE,
  statistics = TRUE,
  mycols = NULL,
  paired = NULL,
  combination = NULL,
  shapes = NULL,
  ID = NULL,
  facet = NULL,
  name = NULL,
  addnumber = TRUE,
  height,
  width,
  strata_var = NULL
)
}
\arguments{
\item{psIN}{Phyloseq object containing microbiome data.}

\item{cate.vars}{Categorical variables to be used for grouping in the plot.}

\item{project}{Name of the project for labeling.}

\item{orders}{Ordering of factors in the plot.}

\item{distance_metrics}{Vector of distance metrics to be used.}

\item{cate.conf}{Additional confounder variables, if any.}

\item{plot}{Type of plot to generate, e.g., "PCoA".}

\item{ellipse}{Logical indicating whether to include ellipses.}

\item{statistics}{Logical indicating whether to include statistical tests.}

\item{mycols}{Custom color palette.}

\item{paired}{Variable for paired analysis, if applicable.}

\item{combination}{Number of groups to combine for comparison.}

\item{shapes}{Shape of points in the plot.}

\item{ID}{Identifier for labeling points.}

\item{facet}{Faceting variable for the plot.}

\item{name}{Optional name for the output plot.}

\item{addnumber}{Logical indicating whether to add sample numbers to groups.}

\item{height}{Height of the plot.}

\item{width}{Width of the plot.}

\item{strata_var}{(NEW) Column name to use as permutation blocks for PERMANOVA.}
}
\value{
PDF(s) and CSV tables on disk.
}
\description{
Generate Biodiversity Plots for Microbiome Data
}
