% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_boxplot_markdown.R
\name{Go_boxplot_markdown}
\alias{Go_boxplot_markdown}
\title{Create Customized Boxplots for Given Categories and Outcomes}
\usage{
Go_boxplot_markdown(
  df,
  cate.vars,
  project,
  outcomes,
  orders = NULL,
  mycols = NULL,
  combination = NULL,
  ylim = NULL,
  title = NULL,
  facet = NULL,
  paired = NULL,
  name = NULL,
  addnumber = TRUE,
  standardsize = TRUE,
  statistics = "yes",
  parametric = "no",
  star = "no",
  xangle = 90,
  font = 10,
  cutoff = 0.1,
  plotCols,
  plotRows
)
}
\arguments{
\item{df}{The data frame containing the variables and outcomes for the boxplots.}

\item{cate.vars}{A vector of column names in 'df' that represent categorical variables for boxplot grouping.}

\item{project}{A character string representing the project name or identifier.}

\item{outcomes}{A vector of column names in 'df' that represent outcomes to be plotted.}

\item{orders}{Optional ordering of the factor levels for the categorical variables.}

\item{mycols}{Optional custom color vector for the boxplots.}

\item{combination}{Optional number of group combinations to compare in the boxplots.}

\item{ylim}{Optional y-axis limits for the boxplots.}

\item{title}{Optional title for the boxplots.}

\item{facet}{Optional column name in 'df' for faceting the boxplots.}

\item{paired}{Optional column name in 'df' for creating paired boxplots.}

\item{name}{Optional name for saving the output plots.}

\item{addnumber}{Logical indicating whether to add sample size to group labels.}

\item{standardsize}{Logical indicating whether to use standard size ratios for the plots.}

\item{statistics}{A character string ("yes" or "no") indicating whether to perform statistical tests.}

\item{parametric}{A character string ("yes" or "no") indicating whether to use parametric tests.}

\item{star}{A character string ("yes" or "no") indicating whether to use star annotations for significance.}

\item{xangle}{Numeric value for the angle of x-axis labels.}

\item{font}{Numeric value specifying the font size for text elements in the plot.}

\item{cutoff}{Numeric value specifying the significance cutoff for statistical annotations.}

\item{plotCols}{Numeric value specifying the number of columns in the multiplot layout.}

\item{plotRows}{Numeric value specifying the number of rows in the multiplot layout.}
}
\value{
The function generates boxplots and optionally saves them but does not return a value.
}
\description{
This function generates boxplots for specified categorical variables and outcomes from a data frame.
It includes options for ordering, color customization, combination of groups, statistical annotations,
and layout adjustments.
}
\examples{
# Assuming 'df' is a data frame with appropriate data
Go_boxplot_markdown(df, cate.vars = c("Group"), project = "MyProject", outcomes = c("Outcome1"),
                    orders = NULL, mycols = c("#00BFC4", "#F8766D"), combination = NULL,
                    ylim = c(0, 10), title = "Boxplot Example", facet = NULL, paired = NULL,
                    name = "Boxplot", addnumber = TRUE, standardsize = TRUE, statistics = "yes",
                    parametric = "no", star = "no", xangle = 90, font = 10, cutoff = 0.1,
                    plotCols = 2, plotRows = 1)

}
