% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_clme_V5.R
\name{Go_clme}
\alias{Go_clme}
\title{Conduct Conditional Linear Mixed Effects (CLME) Analysis on Phyloseq Data}
\usage{
Go_clme(
  psIN,
  cate.vars,
  project,
  paired,
  mycols = NULL,
  addnumber = TRUE,
  standardsize = TRUE,
  node,
  decreasing,
  timepoint,
  ID,
  orders,
  xangle,
  name,
  height,
  width,
  plotCols,
  plotRows
)
}
\arguments{
\item{psIN}{Phyloseq object containing the data for analysis.}

\item{cate.vars}{Categorical variables used in the analysis.}

\item{project}{Name of the project or analysis.}

\item{paired}{Logical value indicating if the data is paired.}

\item{mycols}{Color specifications for the plot.}

\item{addnumber}{Logical value to add the number of samples in each group.}

\item{standardsize}{Logical value to standardize plot size.}

\item{node}{Node parameter for umbrella pattern in CLME.}

\item{decreasing}{Logical value indicating the decreasing trend in CLME.}

\item{timepoint}{Time point variable in the data.}

\item{ID}{Identifier for individual samples.}

\item{orders}{Order of levels in the categorical variables.}

\item{xangle}{Angle of x-axis labels in the plot.}

\item{name}{Optional name for the analysis.}

\item{height}{Height of the output plot.}

\item{width}{Width of the output plot.}

\item{plotCols}{Number of columns for multiple plots layout.}

\item{plotRows}{Number of rows for multiple plots layout.}
}
\value{
Generates CLME plots as PDF files and saves them in a specified directory.
}
\description{
This function performs a conditional linear mixed effects analysis on data from a Phyloseq object.
It supports various customization options including grouping, node specification, and plotting parameters.
}
\details{
The function applies a CLME model to the Phyloseq data, considering various parameters and
constraints. It can handle paired data, apply node constraints, and consider decreasing trends.
The function generates plots for each categorical variable and subgroup within the data.
}
\examples{
# psIN is a Phyloseq object
# Example usage:
Go_clme(psIN = psIN,
        cate.vars = c("Treatment", "Condition"),
        project = "MyProject",
        paired = TRUE,
        node = 10,
        decreasing = TRUE,
        timepoint = "Day",
        ID = "PatientID",
        orders = c("Day1", "Day2", "Day3"),
        xangle = 45,
        name = "Analysis1",
        height = 10,
        width = 10,
        plotCols = 2,
        plotRows = 2)

}
