% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_DA_v7.R
\name{Go_DA}
\alias{Go_DA}
\title{Differential Abundance Analysis Using DESeq2 and ANCOM-BC}
\usage{
Go_DA(
  psIN,
  project,
  filter,
  taxanames = NULL,
  data_type = "other",
  cate.vars,
  cate.conf = NULL,
  cont.conf = NULL,
  orders = NULL,
  name = NULL,
  fdr = 0.05
)
}
\arguments{
\item{psIN}{The phyloseq object containing OTU data.}

\item{project}{A string representing the name of the project.}

\item{filter}{The minimum count threshold for filtering OTUs.}

\item{taxanames}{Optional, the taxonomic level for aggregation.}

\item{data_type}{Type of data ('taxonomy', 'kegg', 'pathway', 'RNAseq', or 'other').}

\item{cate.vars}{A vector of categorical variables for DA analysis.}

\item{cate.conf}{Optional, categorical confounders for the model.}

\item{cont.conf}{Optional, continuous confounders for the model.}

\item{orders}{Optional, order of factor levels for the categorical variable.}

\item{name}{Optional, a name for the analysis.}

\item{fdr}{The false discovery rate threshold for significance.}
}
\value{
The function outputs CSV files and RDS files of significant taxa and associated DA results, including DESeq2 and ANCOM-BC analysis.
}
\description{
This function performs differential abundance (DA) analysis on a phyloseq object.
}
\details{
The function first aggregates taxa at the specified level (if provided), filters the data based on the given cutoff, and then applies DESeq2 and optionally ANCOM-BC for DA analysis, considering specified confounders.
}
\examples{
# Example usage:
Go_DA(psIN = my_phyloseq_object, project = "MyStudy", filter = 0.00005, taxanames = "Genus", data_type = "taxonomy", cate.vars = c("Treatment", "Condition"), cate.conf = c("Age", "Gender"), cont.conf = "BMI", orders = c("Control", "Treatment"), name = "DA_Analysis", fdr = 0.05)

}
