% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_DA_heatmap_v1.R
\name{Go_DA_heat}
\alias{Go_DA_heat}
\title{Create a Differential Abundance Heatmap}
\usage{
Go_DA_heat(
  df,
  project,
  data_type,
  font,
  addnumber = TRUE,
  facet = NULL,
  pval,
  fc,
  orders,
  name,
  height,
  width
)
}
\arguments{
\item{df}{Data frame containing differential abundance results.}

\item{project}{Name of the project.}

\item{data_type}{Type of data used, e.g., 'dada2' or 'other'.}

\item{font}{Font size for the plot text.}

\item{addnumber}{Logical indicating whether to add numbers to the plot.}

\item{facet}{Name of the facet column in the data frame.}

\item{fc}{Fold Change threshold for filtering significant features.}

\item{orders}{Vector of ordered factor levels for arranging groups.}

\item{name}{Optional name for the saved plot.}

\item{height}{Height of the output plot.}

\item{width}{Width of the output plot.}

\item{groupby}{Name of the grouping column in the data frame.}

\item{fdr}{False Discovery Rate threshold for filtering significant features.}
}
\value{
A heatmap plot saved as a PDF file, showing differentially abundant features.
}
\description{
Create a Differential Abundance Heatmap
}
\details{
This function generates a heatmap to visualize differential abundance results. It highlights significant features based on FDR and fold change thresholds and arranges them according to specified orders. The function is designed for comparative analysis in microbiome studies.
}
\examples{
Go_DA_heat(df = diff_abundance_results,
           project = "MyProject",
           data_type = "dada2",
           facet = "Condition",
           groupby = "Group",
           font = 12,
           addnumber = TRUE,
           fdr = 0.05,
           fc = 2,
           orders = c("Group1", "Group2"),
           name = "HeatmapPlot",
           height = 10,
           width = 15)

}
