% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_Deseq2_markdown.R
\name{Go_Deseq2_markdown}
\alias{Go_Deseq2_markdown}
\title{Differential Abundance Analysis Using DESeq2}
\usage{
Go_Deseq2_markdown(
  psIN,
  project,
  taxanames = NULL,
  cate.outs,
  cate.conf = NULL,
  cont.conf = NULL,
  orders = NULL,
  name = NULL
)
}
\arguments{
\item{psIN}{Phyloseq object containing OTU/ASV counts and associated sample data.}

\item{project}{Name or identifier for the project or dataset being analyzed.}

\item{taxanames}{Optional argument to aggregate taxa at a specified taxonomic level (e.g., "Genus", "Family").}

\item{cate.outs}{Vector of column names in the sample data of 'psIN' representing categorical outcome variables.}

\item{cate.conf}{Optional vector of column names in the sample data of 'psIN' representing categorical confounding variables.}

\item{cont.conf}{Optional vector of column names in the sample data of 'psIN' representing continuous confounding variables.}

\item{orders}{Optional ordering for the levels of the categorical variables.}

\item{name}{Optional name to assign to the results (useful for identification if storing multiple result sets).}
}
\value{
A list of data frames, where each data frame contains the results of the differential abundance analysis
for one outcome variable. Includes statistics such as log2 fold changes, p-values, and adjusted p-values,
along with taxonomic information for each feature.
}
\description{
This function performs differential abundance analysis using DESeq2 on data contained within a phyloseq object.
It can handle categorical outcomes and optionally accommodate confounding variables. The function also
allows for aggregation at different taxonomic levels.
}
\examples{
# Assuming 'ps' is a phyloseq object with appropriate data
results <- Go_Deseq2_markdown(ps, "my_project",
                             cate.outs = c("Group"),
                             cate.conf = c("Gender"),
                             cont.conf = NULL,
                             orders = NULL,
                             name = "Analysis1")

}
