% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_deseq2fishtaco.R
\name{Go_deseq2fishtaco}
\alias{Go_deseq2fishtaco}
\title{Convert DESeq2 Results for FishTaco Analysis}
\usage{
Go_deseq2fishtaco(psIN, project, file_path)
}
\arguments{
\item{psIN}{The Phyloseq object containing the microbiome data.}

\item{project}{The name of the project, used for naming output files.}

\item{file_path}{Path to the folder containing DESeq2 output files.}
}
\value{
For each DESeq2 output file, a corresponding CSV file is created in the specified output directory. These files are formatted for use with the FishTaco tool.
}
\description{
Convert DESeq2 Results for FishTaco Analysis
}
\details{
This function processes DESeq2 results and prepares them for FishTaco analysis. It reads the DESeq2 output files, extracts relevant taxonomic information from the Phyloseq object, and aggregates data at the KEGG Orthology (KO) level. The function then filters and exports the processed data for each DESeq2 file, making it compatible with FishTaco input requirements.
}
\examples{
Go_deseq2fishtaco(psIN = my_phyloseq_object,
                  project = "MyMicrobiomeProject",
                  file_path = "path/to/deseq2/results")

}
