% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_dirInfo.R
\name{Go_dirInfo}
\alias{Go_dirInfo}
\title{Set Working Directory and Return Paths for Image Storage}
\usage{
Go_dirInfo(
  Current_working_dir = NA,
  Image_locations = NA,
  DA_image_location = NA
)
}
\arguments{
\item{Image_locations}{A subdirectory within the current working directory
where .pdf images are stored. This is optional.}

\item{DA_image_location}{A subdirectory within the current working directory
where .pdf images for differential abundance tests are stored. This is optional.}

\item{Current_woringing_dir}{A character string specifying the path to set
as the current working directory. If \code{NA}, the function will output
the current working directory and its contents.}
}
\value{
A list containing the paths for the current working directory,
image directory, and differential abundance test image directory.
If \code{Corrent_woringing_dir} is \code{NA}, no list is returned and
function outputs directory information directly.
}
\description{
This function sets the working directory to a specified path and returns the
paths for storing .pdf images and differential abundance test images.
If no directory is specified, it displays the current working directory,
lists its contents, and describes the usage of other parameters.
}
\examples{
\dontrun{
  dir_info <- Go_dirInfo("path/to/directory", "images", "DA_images")
  print(dir_info)

  # If Corrent_woringing_dir is NA
  Go_dirInfo()
}

}
