% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_dist_V4.R
\name{Go_dist}
\alias{Go_dist}
\title{Calculate Distance Matrices for Phyloseq Object}
\usage{
Go_dist(psIN, project, cate.vars, name = NULL, distance_metrics)
}
\arguments{
\item{psIN}{A Phyloseq object to calculate the distance matrices for.}

\item{project}{A string indicating the project name for file naming.}

\item{cate.vars}{An array of categorical variables to consider in the distance matrix calculations.}

\item{name}{(Optional) A string to add as a suffix to the file name.}

\item{distance_metrics}{An array of distance metrics to use for calculations.}
}
\value{
A list of distance matrices for each specified metric.
}
\description{
This function calculates distance matrices for a given Phyloseq object using various distance metrics.
}
\details{
The function calculates distance matrices using specified distance metrics and saves them in specified directories.
Supported distance metrics include unifrac, wunifrac, dpcoa, jsd, manhattan, euclidean, canberra, bray, kulczynski, and jaccard.
It also handles subsetting for different categories and calculates distances both within and across different groups.
}
\examples{
Go_dist(psIN = my_phyloseq_object,
        project = "MyProject",
        cate.vars = c("Variable1", "Variable2"),
        name = "Analysis1",
        distance_metrics = c("bray", "unifrac"))

}
