% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_dist_markdown.R
\name{Go_dist_markdown}
\alias{Go_dist_markdown}
\title{Calculate Distances for Microbiome Data}
\usage{
Go_dist_markdown(psIN, project, cate.vars, name = NULL, distance_metrics)
}
\arguments{
\item{psIN}{A Phyloseq object containing the OTU/ASV counts and associated sample data.}

\item{project}{Name or identifier for the project or dataset being analyzed.}

\item{cate.vars}{Vector of column names in the sample data of 'psIN' representing categorical variables for distance calculation.}

\item{name}{Optional name for the analysis output.}

\item{distance_metrics}{Vector of distance metrics to be used for the calculations, such as "bray", "unifrac", etc.}
}
\value{
A list of distance matrices, one for each specified distance metric.
}
\description{
This function calculates distances for microbiome data in a Phyloseq object using specified distance metrics.
It supports multiple categorical variables and distance metrics.
}
\examples{
# Assuming 'ps' is a phyloseq object with appropriate data
Go_dist_markdown(ps, "my_project", c("Group"), distance_metrics = c("bray", "unifrac"))

}
