% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_emptyMap.R
\name{Go_emptyMap}
\alias{Go_emptyMap}
\title{Create Empty Mapping and Metadata Tables}
\usage{
Go_emptyMap(psIN, project)
}
\arguments{
\item{psIN}{The phyloseq object for which the mapping and metadata tables are to be created.}

\item{project}{A string representing the name of the project.}
}
\value{
Generates CSV files with empty mapping and metadata tables, saved in a specified directory.
}
\description{
This function generates empty mapping and metadata tables for a given phyloseq object and project.
}
\details{
The function creates two CSV files:
\itemize{
\item An empty mapping table with columns like \code{SampleID}, \code{StudyID}, \code{TreatmentGroup}, \code{Timepoint}, etc.
\item An empty metadata table with columns like \code{StudyID}, \code{Variation1}, \code{Variation2}, etc., and rows for different analyses (e.g., \code{Go_overview}, \code{Go_ancombc}).
The purpose is to provide templates for users to fill in their specific study details and analysis parameters.
}
}
\examples{
# Example usage:
Go_emptyMap(psIN = my_phyloseq_object, project = "MyMicrobiomeProject")

}
