% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_filter.R
\name{Go_filter}
\alias{Go_filter}
\title{Filter Low Abundance Taxa from Phyloseq Object}
\usage{
Go_filter(psIN, cutoff)
}
\arguments{
\item{psIN}{Phyloseq object to be filtered.}

\item{cutoff}{Abundance threshold for filtering taxa.}
}
\value{
A filtered Phyloseq object.
}
\description{
Filter Low Abundance Taxa from Phyloseq Object
}
\details{
This function filters out low-abundance taxa from a Phyloseq object based on a specified abundance threshold. It first removes samples with zero or very low total counts, then applies a relative abundance filter to exclude taxa below the cutoff threshold.
}
\examples{
ps_filtered <- Go_filter(psIN = my_phyloseq_object, cutoff = 0.00005)

}
