% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_filter_markdown.R
\name{Go_filter_markdown}
\alias{Go_filter_markdown}
\title{Filter Phyloseq Data}
\usage{
Go_filter_markdown(psIN, cutoff)
}
\arguments{
\item{psIN}{A Phyloseq object containing microbiome data.}

\item{cutoff}{A numeric value specifying the minimum average relative abundance threshold for taxa filtering.}
}
\value{
A Phyloseq object after applying the filtering criteria.
}
\description{
This function filters a Phyloseq object to remove low-abundance taxa based on a specified cutoff.
It removes ASVs (Amplicon Sequence Variants) with zero counts and taxa below the cutoff threshold.
}
\examples{
# Assuming 'ps' is a Phyloseq object
filtered_ps <- Go_filter_markdown(ps, 0.00005)

}
