% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_forestplot.R
\name{Go_forestplot}
\alias{Go_forestplot}
\title{Go_forestplot}
\usage{
Go_forestplot(df, project, outcomes, formular, name = NULL, height, width)
}
\arguments{
\item{df}{A data frame containing the data to be used for the models.}

\item{project}{A character string specifying the project name. Used for naming output directories and files.}

\item{outcomes}{A character vector specifying the names of the outcome variables to be modeled.}

\item{formular}{A character string representing the formula for the model. It should contain a placeholder ("\%s") for the outcome variable.}

\item{name}{An optional character string to include in the output file names. Default is NULL.}

\item{height}{A numeric value specifying the height of the output plots in inches.}

\item{width}{A numeric value specifying the width of the output plots in inches.}
}
\value{
This function does not return a value. It creates and saves forest plot PDFs in the specified directory.
}
\description{
This function generates forest plots for linear mixed models (LMM) based on the given outcomes and formula. It saves the plots as PDF files in an output directory.
}
\examples{
\dontrun{
df <- your_data_frame
outcomes <- c("outcome1", "outcome2")
formular <- "\%s ~ predictor1 + predictor2 + (1|random_effect)"
Go_forestplot(df, project = "MyProject", outcomes = outcomes, formular = formular, height = 6, width = 8)
}
}
