% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_function2ps_V2_humann3.R
\name{Go_function2ps_test}
\alias{Go_function2ps_test}
\title{Convert Functional Annotation Data to Phyloseq Object}
\usage{
Go_function2ps_test(tabPath, project = NULL, func.type, name = NULL)
}
\arguments{
\item{tabPath}{The path to the functional annotation table file.}

\item{project}{Optional project name for labeling purposes.}

\item{func.type}{The type of functional annotation, either PICRUSt or HUMAnN.}

\item{name}{Optional name for the output object.}
}
\value{
A Phyloseq object containing the functional annotation data, ready for ecological and bioinformatic analyses.
}
\description{
Convert Functional Annotation Data to Phyloseq Object
}
\details{
This function converts a functional annotation table into a Phyloseq object for further analysis in ecological and bioinformatic contexts. It supports both PICRUSt and HUMAnN annotation types. The function reads the annotation table, cleans it by removing certain unwanted categories, and structures it into a format compatible with Phyloseq.
}
\examples{
ps_object <- Go_function2ps(tabPath = "path/to/annotation_table.csv",
                            project = "MyProject",
                            func.type = "PICRUSt",
                            name = "FunctionalAnalysis")

}
