% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_Groupheatmap.R
\name{Go_Groupheatmap}
\alias{Go_Groupheatmap}
\title{Go_Groupheatmap}
\usage{
Go_Groupheatmap(
  df,
  SampleData,
  project,
  Group,
  orders,
  title = NULL,
  mycols = NULL,
  name = NULL,
  top_n = 30,
  normalization = "log",
  width = 16,
  height = 8,
  x_label = NULL
)
}
\arguments{
\item{df}{A data frame containing the data to be visualized. Rows represent features, and columns represent samples.}

\item{SampleData}{A data frame with sample metadata. Row names should match the column names of \code{df}.}

\item{project}{A character string specifying the project name, used for file naming.}

\item{Group}{A character string specifying the column name in \code{SampleData} that contains group information for the samples.}

\item{orders}{A character vector specifying the order of groups to display on the heatmap. If NULL, groups are displayed in descending order of their sizes.}

\item{title}{A character string specifying the title of the heatmap.}

\item{name}{An optional character string for additional naming in the output file.}

\item{top_n}{An integer specifying the number of top features to include in the heatmap based on mean depth. Default is 30. If NULL, all features are included.}

\item{normalization}{A character string specifying the normalization method. Options are "log" (log transformation and scaling), "Zscore" (z-score normalization), or "none" (no normalization). Default is "log".}

\item{width}{An integer specifying the width of the output PDF file in inches. Default is 16.}

\item{height}{An integer specifying the height of the output PDF file in inches. Default is 8.}

\item{x_label}{An optional character string specifying the column name in \code{SampleData} to be used as x-axis labels in the heatmap. If NULL, the group information is used as labels.}
}
\value{
The function saves a PDF file with the heatmap and returns nothing.
}
\description{
This function generates a heatmap for the top \code{n} features of a data matrix, grouped by a specified category. It allows for normalization of the data and customization of plot appearance.
}
\details{
The function performs the following steps:
\itemize{
\item Selects the top \code{n} features based on mean depth or uses all features if \code{top_n} is NULL.
\item Reorders the data based on common sample IDs.
\item Applies the specified normalization method.
\item Generates a heatmap and saves it as a PDF file with the specified dimensions.
}
}
\examples{
\dontrun{
df <- your_data_matrix
SampleData <- your_sample_metadata
Go_Groupheatmap(df, SampleData, project = "Project1", Group = "Group", orders = c("Group1", "Group2"), title = "Heatmap Example", top_n = 20, normalization = "log", width = 12, height = 10)
}
}
