% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_hmPath2ps_V1.R
\name{Go_hmPath2ps}
\alias{Go_hmPath2ps}
\title{Create Phyloseq Object from Pathway Data}
\usage{
Go_hmPath2ps(project, tool, pathway, name, alpha)
}
\arguments{
\item{project}{Name of the project associated with the data.}

\item{tool}{Tool used for generating pathway data (e.g., 'humann2', 'picrust2').}

\item{pathway}{Path to the pathway data file.}

\item{name}{Optional name for the output file.}

\item{alpha}{Threshold value for filtering low-abundance pathways.}
}
\value{
A Phyloseq object containing the processed pathway data, which is also saved as an RDS file.
}
\description{
Create Phyloseq Object from Pathway Data
}
\details{
This function processes pathway data generated by tools like HUMAnN2 or PICRUSt2, creating a Phyloseq object for downstream analysis. It includes options for filtering low-abundance pathways and extracting relevant information for creating taxonomic and abundance tables.
}
\examples{
Go_hmPath2ps(project = "MyProject",
             tool = "humann2",
             pathway = "pathway_data.tsv",
             name = "pathway_analysis",
             alpha = 50)

}
