% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_human2ps_V3.R
\name{Go_huamnn2ps}
\alias{Go_huamnn2ps}
\title{Convert Functional Data to Phyloseq Object}
\usage{
Go_huamnn2ps(project, tool, koTOpath, koSta, name, alpha)
}
\arguments{
\item{project}{Name of the project.}

\item{tool}{Name of the tool used (e.g., 'humann2' or 'picrust2').}

\item{koTOpath}{Path to the file containing KEGG Orthology to Pathway mappings.}

\item{koSta}{Path to the file containing KEGG Orthology stratified data.}

\item{name}{Optional name for the saved Phyloseq object.}

\item{alpha}{Threshold value for filtering data based on row sums.}
}
\value{
A Phyloseq object representing the processed functional data. This object is also saved as an RDS file.
}
\description{
Convert Functional Data to Phyloseq Object
}
\details{
This function processes the output of functional prediction tools like HUMAnN2 or PICRUSt2 and converts it into a Phyloseq object. It includes steps for data cleaning, subsetting, and reformatting.
}
\examples{
Go_huamnn2ps(project = "MyMicrobiomeProject",
             tool = "humann2",
             koTOpath = "KO_to_Pathway.csv",
             koSta = "KO_Stratified.csv",
             name = "FunctionalData",
             alpha = 50)

}
