% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_human2ps_stratified.R
\name{Go_huamnn2ps_stratified}
\alias{Go_huamnn2ps_stratified}
\title{Stratified Conversion of KEGG Orthology to Phyloseq Object}
\usage{
Go_huamnn2ps_stratified(project, koTOpath, koSta, alpha, name)
}
\arguments{
\item{project}{A string representing the project name, used for file naming and directory creation.}

\item{koTOpath}{The file path for the KEGG Orthology to Pathway mapping data.}

\item{koSta}{The file path for the KEGG Orthology stratified data.}

\item{alpha}{A threshold for filtering pathways based on read count.}

\item{name}{(Optional) A name for labeling the output.}
}
\value{
Generates a phyloseq object from stratified KEGG Orthology data and saves it as an RDS file.
}
\description{
This function converts KEGG Orthology data into a phyloseq object with stratification and filtering.
}
\details{
The function reads KEGG Orthology data, applies filtering based on the provided threshold, and
converts it into a phyloseq object. It includes additional information about the pathways and
stratifies the data for in-depth analysis.
}
\examples{
# Example usage:
Go_huamnn2ps_stratified(project = "MyProject",
                       koTOpath = "path_to_ko_to_path.csv",
                       koSta = "path_to_ko_stratified.csv",
                       alpha = 50,
                       name = "MyAnalysis")

}
