% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_krakenLog.R
\name{Go_krakenLog}
\alias{Go_krakenLog}
\title{Process Kraken Output for Microbiome Data}
\usage{
Go_krakenLog(project, hm.log, st.log, mpa)
}
\arguments{
\item{project}{Name of the project associated with the data.}

\item{hm.log}{Path to the human read log file from Kraken analysis.}

\item{st.log}{Path to the standard read log file from Kraken analysis.}

\item{mpa}{Path to the Kraken multi-pathway analysis (mpa) output file.}
}
\value{
A consolidated data frame containing processed Kraken log information. This data frame is also saved as a CSV file.
}
\description{
Process Kraken Output for Microbiome Data
}
\details{
This function processes and merges log files generated by Kraken, a popular tool for metagenomic analysis. It combines information about human reads, standard reads, and multi-pathway analysis into a comprehensive table.
}
\examples{
Go_krakenLog(project = "MyMicrobiomeProject",
             hm.log = "human_reads_log.txt",
             st.log = "standard_reads_log.txt",
             mpa = "kraken_mpa_output.txt")

}
