% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_krakenSummaryPlot.R
\name{Go_krakenSummaryPlot}
\alias{Go_krakenSummaryPlot}
\title{Create Kraken Summary Plot}
\usage{
Go_krakenSummaryPlot(
  project,
  logTab,
  name = NULL,
  with_human = T,
  legend = "right",
  height,
  width
)
}
\arguments{
\item{project}{A string indicating the project name for file naming.}

\item{logTab}{The path to the log file containing Kraken analysis results.}

\item{name}{(Optional) A string to add as a suffix to the file name.}

\item{with_human}{A logical value indicating whether to include human reads in the plot (default is TRUE).}

\item{legend}{A string indicating the position of the legend ("right" or other valid ggplot2 positions).}

\item{height}{The height of the output plot.}

\item{width}{The width of the output plot.}
}
\value{
A bar plot saved as a PDF file in the specified output directory.
}
\description{
This function creates a bar plot summarizing the results of a Kraken analysis.
}
\details{
The function reads Kraken analysis results and prepares a bar plot showing the relative abundance of various
categories like Archaea, Bacteria, Fungi, Viruses, and optionally human reads. The plot is saved as a PDF.
}
\examples{
Go_krakenSummaryPlot(project = "MyProject",
                     logTab = "path/to/kraken_log.csv",
                     name = "Analysis1",
                     with_human = TRUE,
                     legend = "right",
                     height = 6,
                     width = 8)

}
