% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_krakenTops.R
\name{Go_krakenTops}
\alias{Go_krakenTops}
\title{Process Kraken MPA Output for Top Taxa and Prepare for RPKM Analysis}
\usage{
Go_krakenTops(project, mpa)
}
\arguments{
\item{project}{Name of the project or analysis.}

\item{mpa}{File path of the Kraken MPA output.}
}
\value{
The function processes the Kraken MPA data, extracts the top taxa, and saves the relevant
information for further RPKM analysis. It saves a Phyloseq object and other required data as .csv and .rds files.
}
\description{
This function processes Kraken MPA output, focusing on top taxa, and prepares data for RPKM analysis.
}
\details{
The function performs a series of steps to process Kraken MPA output. It reads the MPA file,
extracts top taxa information, formats the data, and saves it for downstream RPKM analysis.
It generates a Phyloseq object and exports relevant tables for species names and OTU.
}
\examples{
# Example usage:
Go_krakenTops(project = "MyProject",
              mpa = "path/to/kraken_mpa_output.txt")

}
