% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_linear_V4.R
\name{Go_linear}
\alias{Go_linear}
\title{Linear Regression Analysis for Continuous Variables}
\usage{
Go_linear(
  df,
  cont.vars,
  project,
  outcomes,
  method = "lm",
  mycols = NULL,
  maingroup = NULL,
  orders = NULL,
  name = NULL,
  height,
  width,
  plotCols,
  plotRows
)
}
\arguments{
\item{df}{A data frame containing the variables for analysis.}

\item{cont.vars}{A vector of continuous variable names to be analyzed.}

\item{project}{A string representing the project name, used for file naming and directory creation.}

\item{outcomes}{A vector of outcome variable names to be analyzed.}

\item{method}{The regression method to use, with options 'glm', 'lm', 'loess', or 'gam'.}

\item{mycols}{(Optional) A vector of colors for plotting.}

\item{maingroup}{(Optional) The main grouping variable for stratification.}

\item{orders}{(Optional) The order of factor levels for plotting.}

\item{name}{(Optional) A name for labeling plots.}

\item{height}{Plot height.}

\item{width}{Plot width.}

\item{plotCols}{The number of columns for the plot layout.}

\item{plotRows}{The number of rows for the plot layout.}
}
\value{
Generates plots of the regression analyses and saves them as a PDF.
}
\description{
This function performs linear regression analysis on continuous variables
in a given data frame, optionally stratified by a main group.
}
\details{
The function provides an easy way to perform and visualize linear regression
analyses on multiple continuous variables against specified outcomes. It
supports various regression methods and allows stratification by a main group.
}
\examples{
# Example usage:
Go_linear(df = my_data_frame,
          cont.vars = c("var1", "var2"),
          project = "MyProject",
          outcomes = c("outcome1", "outcome2"),
          method = "lm",
          height = 6,
          width = 8,
          plotCols = 2,
          plotRows = 2)

}
