% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_mergeTab_v1.R
\name{Go_mergeTab}
\alias{Go_mergeTab}
\title{Merge Multiple Data Tables Based on a Common Pattern}
\usage{
Go_mergeTab(pattern, file_path)
}
\arguments{
\item{pattern}{The pattern used to identify and match files for merging.}

\item{file_path}{The directory path where the files to be merged are located.}
}
\value{
A single data frame consisting of the row-wise merged contents of all files matching the specified pattern in the given directory.
}
\description{
Merge Multiple Data Tables Based on a Common Pattern
}
\details{
This function aggregates multiple data tables into a single data frame. It searches for files in a specified directory that match a given pattern and merges them row-wise. This function is particularly useful for consolidating similar data spread across multiple files, such as results from repetitive experiments or outputs from batch processes.
}
\examples{
merged_data <- Go_mergeTab(pattern = ".csv",
                           file_path = "path/to/data/files")

}
