% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_mpaTops.R
\name{Go_mpaTops}
\alias{Go_mpaTops}
\title{Process Kraken and metaphlan MPA Output for Top Taxa}
\usage{
Go_mpaTops(project, mpa, kingdom = "d__Bacteria")
}
\arguments{
\item{project}{Name of the project or analysis.}

\item{mpa}{File path of the Kraken and metaphlan MPA output.}
}
\value{
The function processes the Kraken MPA data, extracts the top taxa, and saves the relevant
information for further RPKM analysis. It saves a Phyloseq object and other required data as .csv and .rds files.
}
\description{
This function processes Kraken and metaphlan  MPA output, focusing on top taxa.
}
\details{
The function performs a series of steps to process Kraken and metaphlan MPA output. It reads the MPA file,
extracts top taxa information, formats the data, and saves it for downstream RPKM analysis.
It generates a Phyloseq object and exports relevant tables for species names and OTU.
}
\examples{
# Example usage:
Go_krakenTops(project = "MyProject",
              mpa = "path/to/kraken_mpa_output.txt")

}
