% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_path_v1.R
\name{Go_path}
\alias{Go_path}
\title{Create Directories for Project Outputs}
\usage{
Go_path(project, pdf, table, path)
}
\arguments{
\item{project}{The name of the project, used to name the main directory.}

\item{pdf}{Indicator (yes/no) for creating a directory to store PDF files.}

\item{table}{Indicator (yes/no) for creating a directory to store table files.}

\item{path}{Additional specific path for creating another directory.}
}
\value{
A list of paths to the created directories, with keys corresponding to the types of outputs (pdf, table, and other specified paths).
}
\description{
This function creates directories for storing various outputs of a microbiome analysis project.
}
\details{
The function creates a main directory for the project and, based on the input parameters, can create additional subdirectories for:
\itemize{
\item PDF files (\code{pdf} parameter)
\item Table files (\code{table} parameter)
\item Any other specified type of files or outputs (\code{path} parameter)
}

It is useful for organizing different types of outputs in a structured manner within the project.
}
\examples{
# Example usage
}
