% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_perm_V16.R
\name{Go_perm}
\alias{Go_perm}
\title{Global Permutation Analysis for Microbiome Data}
\usage{
Go_perm(
  psIN,
  vars,
  project,
  distance_metrics,
  multi = FALSE,
  name = NULL,
  strata_var = NULL
)
}
\arguments{
\item{psIN}{A phyloseq object containing microbiome data.}

\item{vars}{A vector of categorical variables for which PERMANOVA will be performed.}

\item{project}{A string representing the project name, used for file naming and directory creation.}

\item{distance_metrics}{A vector of distance metrics to be used in the analysis.}

\item{name}{An optional string for naming the output files.}

\item{strata_var}{A variable used to define strata for the PERMANOVA analysis, ensuring that permutations are constrained within each stratum (e.g., patient ID).}

\item{mul.vars}{A boolean value indicating whether to use multiple variables in the PERMANOVA model.}
}
\value{
Returns a data frame containing the results of the PERMANOVA analysis, including degrees of freedom, sums of squares, R-squared values, F-model statistics, p-values, and adjusted p-values.
}
\description{
This function performs global permutation (PERMANOVA) analysis on microbiome data.
}
\details{
The function conducts PERMANOVA analysis to understand the impact of various categorical variables on microbial community composition. It supports multiple distance metrics and can handle both single and multiple variables in the model. The function outputs the results in a CSV file and returns them as a data frame.
}
\examples{
# Example usage:
permanova_results <- Go_perm(psIN = my_phyloseq_object,
                             vars = c("TreatmentGroup", "AgeGroup"),
                             project = "MyMicrobiomeStudy",
                             distance_metrics = c("bray", "unifrac"),
                             multi = FALSE,
                             strata_var=NULL,
                             name = "MyAnalysis")

}
