% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_pheatmap_V2.R
\name{Go_pheatmap}
\alias{Go_pheatmap}
\title{Generate Heatmaps from Phyloseq Data}
\usage{
Go_pheatmap(
  psIN,
  project,
  title = NULL,
  group1 = NULL,
  group2 = NULL,
  group3 = NULL,
  group4 = NULL,
  Ntax = NULL,
  name = NULL,
  col_orders = NULL,
  show_rownames = T,
  show_colnames = F,
  cutree_rows = NA,
  cutree_cols = NA,
  cluster_rows = T,
  cluster_cols = T,
  row_gap = NULL,
  column_gap = NULL,
  showPhylum = T,
  width,
  patchwork = FALSE
)
}
\arguments{
\item{psIN}{Phyloseq object containing the data for analysis.}

\item{project}{Name of the project or analysis.}

\item{title}{Title of the heatmap.}

\item{group1}{First group for annotation and coloring in the heatmap.}

\item{group2}{Second group for annotation and coloring in the heatmap.}

\item{group3}{Third group for annotation and coloring in the heatmap.}

\item{group4}{Fourth group for annotation and coloring in the heatmap.}

\item{Ntax}{Number of top taxa to include in the heatmap (if NULL, all taxa are considered).}

\item{name}{Optional name for the analysis.}

\item{col_orders}{Custom order of columns in the heatmap.}

\item{show_rownames}{Logical value to show or hide row names.}

\item{show_colnames}{Logical value to show or hide column names.}

\item{cutree_rows}{Number of clusters for row dendrogram cutting.}

\item{cutree_cols}{Number of clusters for column dendrogram cutting.}

\item{cluster_rows}{Logical value to cluster or not cluster rows.}

\item{cluster_cols}{Logical value to cluster or not cluster columns.}

\item{showPhylum}{Logical value to show or hide phylum-level annotation.}

\item{width}{Width of the output heatmap.}

\item{patchwork}{Logical. If \code{TRUE}, skip saving and return the plot object(s) for use with \code{Gg_patchwork()} or the \pkg{patchwork} package. Default \code{FALSE}.}
}
\value{
Saves heatmap plots as PDF files in a specified directory.
}
\description{
This function creates heatmaps from a Phyloseq object, allowing visualization of
complex microbiome data. It supports grouping, color customization, and hierarchical clustering.
}
\details{
The function preprocesses the Phyloseq object, potentially considering only the top N taxa,
and then generates heatmaps with options for clustering, grouping, and annotation.
It can display phylum-level annotations and supports various customization options.
}
\examples{
# psIN is a Phyloseq object
# Example usage:
Go_pheatmap(psIN = psIN,
            project = "MyProject",
            title = "Sample Heatmap",
            group1 = "Treatment",
            group2 = "Condition",
            Ntax = 50,
            name = "Analysis1",
            col_orders = c("Sample1", "Sample2"),
            show_rownames = TRUE,
            show_colnames = FALSE,
            cluster_rows = TRUE,
            cluster_cols = TRUE,
            showPhylum = TRUE,
            width = 10)

}
