% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_pheatmap_markdown.R
\name{Go_pheatmap_markdown}
\alias{Go_pheatmap_markdown}
\title{Create Heatmap for Microbiome Data}
\usage{
Go_pheatmap_markdown(
  psIN,
  title,
  group1 = NULL,
  group2 = NULL,
  group3 = NULL,
  group4 = NULL,
  Ntax = NULL,
  col_orders = NULL,
  show_rownames = T,
  show_colnames = F,
  cutree_rows = NA,
  cutree_cols = NA,
  cluster_rows = T,
  cluster_cols = T,
  showPhylum = T,
  width
)
}
\arguments{
\item{psIN}{A \code{phyloseq} object containing the microbiome data.}

\item{title}{Title for the heatmap.}

\item{group1, }{group2, group3, group4 Optional grouping variables for columns.}

\item{Ntax}{The number of taxa to include in the heatmap (default is all).}

\item{col_orders}{Optional custom ordering for the columns.}

\item{show_rownames, }{show_colnames Logical indicating whether to show row and column names.}

\item{cutree_rows, }{cutree_cols Hierarchical clustering cut heights for rows and columns.}

\item{cluster_rows, }{cluster_cols Logical indicating whether to cluster rows and columns.}

\item{showPhylum}{Logical indicating whether to show phylum annotations.}

\item{width}{Width of the heatmap.}
}
\value{
The function generates and prints a heatmap but does not return a value.
}
\description{
This function generates a heatmap for microbiome data contained within a \code{phyloseq} object.
It allows for customization of groups, ordering, and display options, and can include
phylogenetic diversity annotations.
}
\examples{
# Assuming 'ps' is a phyloseq object with appropriate data
Go_pheatmap_markdown(ps, title = "My Heatmap", group1 = "SampleType")

}
