% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_biplot_pickTaxa_V1.R
\name{Go_pickTaxa}
\alias{Go_pickTaxa}
\title{Select Best Fitting Taxa for Biplot Analysis}
\usage{
Go_pickTaxa(psIN, project, TaxaLevel, data_type, bestPick)
}
\arguments{
\item{psIN}{A phyloseq object containing microbiome data.}

\item{project}{A string representing the project name, used for file naming.}

\item{TaxaLevel}{A string specifying the taxonomic level to be analyzed.}

\item{data_type}{A string indicating the type of data (e.g., 'dada2', 'Nephele').}

\item{bestPick}{The number of best taxa combinations to be considered.}
}
\value{
Returns a vector of taxa names that represent the best combination for biplot analysis. The function also prints out similarity scores for each combination.
}
\description{
This function identifies the best combination of taxa for biplot analysis based on their similarity scores.
}
\details{
The function aggregates data at the specified taxonomic level and uses bioenv stepwise selection to find the best fitting taxa combinations based on their similarity to environmental variables or other factors. The function is useful for preparing data for biplot analysis in microbiome studies.
}
\examples{
# Example usage:
best_taxa <- Go_pickTaxa(psIN = my_phyloseq_object,
                         project = "MyMicrobiomeStudy",
                         TaxaLevel = "Genus",
                         data_type = "16S",
                         bestPick = 5)

}
