% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_powerCalc.R
\name{Go_powerCalc}
\alias{Go_powerCalc}
\title{Go_powerCalc}
\usage{
Go_powerCalc(effect_size = 1, alpha = 0.05, power = 0.8, n_timepoints = NULL)
}
\arguments{
\item{effect_size}{An integer indicating the effect size to be used in the calculation. Valid values are 1 (small effect), 2 (medium effect), and 3 (large effect). Default is 1.}

\item{alpha}{The significance level for the hypothesis test. Default is 0.05.}

\item{power}{The desired power of the test. Default is 0.8.}

\item{n_timepoints}{An optional integer specifying the number of timepoints in a longitudinal study. If provided, the function calculates power for longitudinal data; otherwise, it calculates power for a two-sample t-test.}
}
\value{
Prints the result of the power calculation, including the required sample size or effect size based on the specified parameters. For longitudinal data, it provides details on timepoints and variance parameters. For a two-sample t-test, it provides the power of the test.
}
\description{
This function calculates the required sample size to achieve a specified power for a study, based on the effect size and significance level. It handles both longitudinal data and two-sample t-tests.
}
\details{
For longitudinal data, the function uses the \code{lmmpower} package to calculate the required sample size, considering intercept variance, slope variance, residual variance, and covariance between intercept and slope.
For a two-sample t-test, the function uses the \code{pwr} package to calculate the required sample size based on the specified effect size, alpha level, and desired power.
}
\examples{
\dontrun{
Go_powerCalc(effect_size = 2, alpha = 0.05, power = 0.8, n_timepoints = 5)
Go_powerCalc(effect_size = 1, alpha = 0.05, power = 0.8)
}
}
