% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_psTotab.R
\name{Go_psTotab}
\alias{Go_psTotab}
\title{Convert Phyloseq Object to Table and Export Sequences}
\usage{
Go_psTotab(psIN, project)
}
\arguments{
\item{psIN}{Input phyloseq object.}

\item{project}{Name of the project for naming output files.}
}
\value{
The function generates two files:
\enumerate{
\item A CSV file containing the abundance data and taxonomy information from the phyloseq object.
\item A FASTA file containing sequence data.
}

Both files are named using the project name and are saved in a specified output directory.
}
\description{
Convert Phyloseq Object to Table and Export Sequences
}
\details{
This function is designed to extract data from a phyloseq object and convert it into a more accessible table format. It also extracts sequence data and saves it as a FASTA file. This is particularly useful for further analysis or data sharing.
}
\examples{
Go_psTotab(psIN = my_phyloseq_object,
           project = "MyMicrobiomeProject")

}
