% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_qq_V2.R
\name{Go_qq}
\alias{Go_qq}
\title{Generate QQ Plots and Histograms for Alpha Diversity Metrics}
\usage{
Go_qq(psIN, project, alpha_metrics, name, height, width)
}
\arguments{
\item{psIN}{Phyloseq object containing microbiome data.}

\item{project}{Name of the project, used for output file naming.}

\item{alpha_metrics}{Vector of alpha diversity metrics to analyze (e.g., "Shannon", "Simpson").}

\item{name}{Optional name to include in the output file's name.}

\item{height}{Height of the output PDF file.}

\item{width}{Width of the output PDF file.}
}
\value{
Generates a PDF file containing QQ plots and histograms for each alpha diversity metric. The file is saved in the specified output directory. Additionally, the function prints the Shapiro-Wilk normality test results for each metric to the console.
}
\description{
Generate QQ Plots and Histograms for Alpha Diversity Metrics
}
\details{
This function creates QQ plots and histograms for various alpha diversity metrics specified in the \code{alpha_metrics} argument. It computes these metrics using the \code{estimate_richness} function from the phyloseq package and then generates corresponding plots to assess the distribution and normality of each metric.
}
\examples{
Go_qq(psIN = my_phyloseq_object,
      project = "MyMicrobiomeProject",
      alpha_metrics = c("Shannon", "Simpson"),
      name = "Analysis1",
      height = 10,
      width = 8)

}
