% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_rare_markdown.R
\name{Go_rare_markdown}
\alias{Go_rare_markdown}
\title{Generate Rarefaction Curves for Microbiome Data}
\usage{
Go_rare_markdown(
  physeq_object,
  step = 10,
  label = NULL,
  color = NULL,
  xlimit,
  plot = TRUE,
  parallel = FALSE,
  se = TRUE
)
}
\arguments{
\item{physeq_object}{A Phyloseq object containing OTU/ASV counts and associated sample data.}

\item{step}{Step size for rarefaction curve plotting.}

\item{label}{Optional labels for the curves.}

\item{color}{Color to be used for the curves.}

\item{xlimit}{Maximum limit for the x-axis (sequence sample size).}

\item{plot}{Logical; if TRUE, the function plots the rarefaction curve.}

\item{parallel}{Logical; if TRUE, uses parallel processing for faster computation.}

\item{se}{Logical; if TRUE, includes standard error in the rarefaction curves.}
}
\value{
A ggplot object representing the rarefaction curve.
This object is returned invisibly and plotted if 'plot' is TRUE.
}
\description{
This function creates rarefaction curves for microbiome data using a Phyloseq object.
It provides an assessment of species richness across varying sequencing depths.
}
\examples{
# Assuming 'ps' is a Phyloseq object
Go_rare_markdown(ps, step = 10, label = "SampleLabel", color = "red", xlimit = 1000, plot = TRUE)

}
