% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_regDA_V8.R
\name{Go_regDA}
\alias{Go_regDA}
\title{Perform Regression Differential Abundance Analysis}
\usage{
Go_regDA(
  psIN,
  cate.outs = NULL,
  con.outs = NULL,
  cate.conf = NULL,
  con.conf = NULL,
  StudyID,
  project,
  pval = 0.05,
  orders,
  nsamps_threshold,
  filt_threshold,
  taxanames,
  name
)
}
\arguments{
\item{psIN}{Phyloseq object containing the data for analysis.}

\item{cate.outs}{Categorical outcomes for regression analysis.}

\item{con.outs}{Continuous outcomes for regression analysis.}

\item{cate.conf}{Confounding categorical variables to adjust in the model.}

\item{con.conf}{Confounding continuous variables to adjust in the model.}

\item{StudyID}{Column name identifying individual study participants or samples.}

\item{project}{Name of the project or analysis.}

\item{pval}{P-value threshold for significance in the analysis.}

\item{orders}{Order of levels in the categorical variables.}

\item{nsamps_threshold}{Threshold for minimum sample abundance.}

\item{filt_threshold}{Threshold for filtering taxa based on sample prevalence.}

\item{taxanames}{Taxonomic rank to aggregate (e.g., "Genus", "Species").}

\item{name}{Optional name for the analysis.}
}
\value{
Generates tables and RDS files for differential abundance results and saves them in specified directories.
}
\description{
This function conducts regression differential abundance analysis on microbiome data using linear mixed-effects models or generalized linear models.
}
\details{
The function applies linear mixed-effects models (LMEMs) or generalized linear models (GLMs) to identify taxa that are differentially abundant across the specified outcomes while adjusting for confounders.
}
\examples{
# Example usage:
Go_regDA(psIN = phyloseq_object,
         cate.outs = c("Group1", "Group2"),
         con.outs = c("Age", "BMI"),
         cate.conf = c("Sex", "Diet"),
         con.conf = c("Time", "Dosage"),
         StudyID = "SubjectID",
         project = "MyProject",
         pval = 0.05,
         orders = c("Control", "Treatment"),
         nsamps_threshold = 0.01,
         filt_threshold = 0.1,
         taxanames = "Genus",
         name = "Analysis1")

}
