% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_regression_V34.R
\name{Go_regression}
\alias{Go_regression}
\title{Perform Regression Analysis on Given Data}
\usage{
Go_regression(
  data,
  project,
  cate.outs = NULL,
  con.outs = NULL,
  cate.vars = NULL,
  con.vars = NULL,
  mul.vars = FALSE,
  randomEff = NULL,
  orders,
  pvalue = 0.05,
  name = NULL
)
}
\arguments{
\item{data}{The data frame containing the variables for analysis.}

\item{project}{The project name for output file naming.}

\item{cate.outs}{Categorical outcome variables (optional).}

\item{con.outs}{Continuous outcome variables (optional).}

\item{cate.vars}{Categorical predictor variables (optional).}

\item{con.vars}{Continuous predictor variables (optional).}

\item{mul.vars}{Flag for multivariate analysis (TRUE/FALSE).}

\item{randomEff}{Random effects for mixed models (optional).}

\item{orders}{Order of factor levels for categorical variables.}

\item{pvalue}{P-value threshold for significance.}

\item{name}{Additional name identifier for output files (optional).}

\item{interaction}{Interaction terms for the model (optional).}
}
\value{
The function generates a CSV file for each outcome variable containing the model coefficients, standard errors, p-values, adjusted p-values, and confidence intervals. It also saves RDS files for multivariate models.
}
\description{
Perform Regression Analysis on Given Data
}
\details{
This function performs various types of regression analysis (linear, logistic, mixed models) based on the specified parameters. It supports univariate, multivariate, and mixed-effects modeling. The function adjusts the p-values using the False Discovery Rate (FDR) method.
}
\examples{
Go_regression(data = my_data,
              project = "MyProject",
              cate.outs = c("Outcome1", "Outcome2"),
              con.outs = "Outcome3",
              cate.vars = c("Predictor1", "Predictor2"),
              con.vars = "Predictor3",
              mul.vars = "Predictor1 + Predictor2 + (1 | Subject)",
              randomEff = "Subject",
              orders = c("Level1", "Level2"),
              pvalue = 0.05,
              name = "Analysis1")

}
