% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_rf_function_sets.R
\name{Go_rfCMplot}
\alias{Go_rfCMplot}
\title{Random Forest Confusion Matrix Plot}
\usage{
Go_rfCMplot(
  rf,
  project,
  trainSetDf,
  testSetDf,
  outcome,
  orders,
  name = NULL,
  height = 2,
  width = 3
)
}
\arguments{
\item{rf}{Random Forest model object.}

\item{project}{Name of the project or analysis.}

\item{trainSetDf}{DataFrame containing the training dataset.}

\item{testSetDf}{DataFrame containing the test dataset.}

\item{outcome}{Name of the outcome variable.}

\item{orders}{Vector of levels for the outcome variable.}

\item{name}{Optional name for the analysis.}

\item{height}{Height of the plot.}

\item{width}{Width of the plot.}
}
\description{
This function generates a confusion matrix plot for a Random Forest model.
}
\details{
The function creates a visual representation of the model's performance on the test dataset using a confusion matrix.
}
\examples{
# Example usage:
Go_rfCMplot(rf = model, project = "MyProject", trainSetDf = trainData, testSetDf = testData, outcome = "Species", orders = c("setosa", "versicolor", "virginica"), name = "Analysis1")

}
