% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_rf_function_sets.R
\name{Go_rfGridSearch}
\alias{Go_rfGridSearch}
\title{Random Forest Grid Search}
\usage{
Go_rfGridSearch(trainSetDf, testSetDf, outcome, orders, formula)
}
\arguments{
\item{trainSetDf}{DataFrame containing the training dataset.}

\item{testSetDf}{DataFrame containing the test dataset.}

\item{outcome}{Name of the outcome variable in the dataset.}

\item{orders}{Vector of levels for the outcome variable.}

\item{formula}{The formula for the Random Forest model.}
}
\value{
A list containing the best mtry, ntree, maxnodes, nodesize, and best_accuracy found during the grid search.
}
\description{
This function performs a grid search to find the best hyperparameters for a Random Forest model.
}
\details{
The function explores combinations of hyperparameters (mtry, ntree, maxnodes, and nodesize) and evaluates the model's performance on the test dataset to find the optimal configuration.
}
\examples{
# Example usage:
Go_rfGridSearch(trainSetDf = trainData, testSetDf = testData, outcome = "Species", orders = c("setosa", "versicolor", "virginica"), formula = Species ~ .)

}
