% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_rf_function_sets.R
\name{Go_rfoobError}
\alias{Go_rfoobError}
\title{Out-of-Bag Error Rate Plot for Random Forest}
\usage{
Go_rfoobError(
  rf,
  project,
  trainSetDf,
  orders,
  outcome,
  mycols = NULL,
  name = NULL,
  height = 2,
  width = 3
)
}
\arguments{
\item{rf}{Random Forest model object.}

\item{project}{Name of the project or analysis.}

\item{trainSetDf}{DataFrame containing the training dataset.}

\item{orders}{Vector of levels for the outcome variable.}

\item{outcome}{Name of the outcome variable.}

\item{mycols}{Colors to use in the plot.}

\item{name}{Optional name for the analysis.}

\item{height}{Height of the plot.}

\item{width}{Width of the plot.}
}
\description{
This function generates a plot of the Out-of-Bag (OOB) error rate across different numbers of trees in a Random Forest model.
}
\details{
The function visualizes how the OOB error rate changes as more trees are added to the model, providing insight into model convergence.
}
\examples{
# Example usage:
Go_rfoobError(rf = model, project = "MyProject", trainSetDf = trainData, orders = c("setosa", "versicolor", "virginica"), outcome = "Species", name = "Analysis1")

}
