% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_rpkmToPs.R
\name{Go_rpkmToPs}
\alias{Go_rpkmToPs}
\title{Convert RPKM Data to Phyloseq Object}
\usage{
Go_rpkmToPs(project, speciesTab, genoemSizeTab, taxaTab)
}
\arguments{
\item{project}{A string indicating the project name used for file naming.}

\item{speciesTab}{A data frame containing species information.}

\item{genoemSizeTab}{A data frame containing genome sizes.}

\item{taxaTab}{A data frame containing taxonomic information.}
}
\value{
A Phyloseq object containing RPKM data structured for microbiome analysis.
}
\description{
This function converts Reads Per Kilobase of transcript, per Million mapped reads (RPKM) data to a Phyloseq object.
}
\details{
The function cleans and processes species names, matches genome sizes, and calculates RPKM values.
It then integrates this information into a Phyloseq object, which is useful for microbiome data analysis.
}
\examples{
Go_rpkmToPs(project = "MyProject",
            speciesTab = species_dataframe,
            genoemSizeTab = genome_size_dataframe,
            taxaTab = taxonomic_dataframe)

}
