% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_SCRubMap.R
\name{Go_SCRubMap}
\alias{Go_SCRubMap}
\title{Go_SCRubMap}
\usage{
Go_SCRubMap(
  SCRubfile,
  IlluSamplesheet,
  IlluLocation,
  Sample_Type,
  map_dir = "3_map"
)
}
\arguments{
\item{SCRubfile}{The filename of the SCRub file located within the specified map directory.}

\item{IlluSamplesheet}{The filename of the Illumina SampleSheet, assumed to be in the same directory as the SCRub file.}

\item{IlluLocation}{The full path to another Illumina SampleSheet containing additional metadata not present in the primary SampleSheet.}

\item{Sample_Type}{Default sample type to be assigned in the augmented data.}

\item{map_dir}{The directory containing the SCRub file and Illumina SampleSheet. Default is "3_map".}
}
\value{
Writes augmented SCRub data to CSV files organized by plate (as determined by index sets) within the specified map directory. Does not return any data structures into the R environment.
}
\description{
Maps SCRub data to Illumina SampleSheet data, augments it with additional metadata, and outputs plate-level files.
}
\details{
The function first checks for the existence of the specified files within the given directory. It then reads and processes the SCRub data, mapping it against the primary Illumina SampleSheet to align sample identifiers and merge relevant metadata. Additional metadata is pulled from a second, more detailed SampleSheet if provided. The function also infers control status and sample type from sample identifiers, and finally, it segments the data by index set and saves it into separate CSV files.

This function is particularly useful in sequencing data processing pipelines where sample identification and metadata integration are required before further analysis.
}
\examples{
\dontrun{
Go_SCRubMap(
  SCRubfile = "sample_scrub.csv",
  IlluSamplesheet = "sample_sheet.csv",
  IlluLocation = "full_sample_sheet.csv",
  Sample_Type = "experimental",
  map_dir = "path/to/map_dir"
)
}
}
