% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_SeqLengths.R
\name{Go_SeqLengths}
\alias{Go_SeqLengths}
\title{Calculate Sequence Lengths for Phyloseq Data}
\usage{
Go_SeqLengths(psIN, filter = NULL)
}
\arguments{
\item{psIN}{A Phyloseq object containing OTU/ASV counts and associated metadata.}

\item{filter}{An optional numeric value specifying the minimum sequence length to include
in the analysis. If NULL (default), no filtering is applied.}
}
\value{
If no filter is applied, it prints and plots the distribution of sequence lengths.
If a filter is applied, it returns a new Phyloseq object containing only sequences
with lengths greater than or equal to the filter value.
}
\description{
This function calculates the sequence lengths from ASV sequences in a Phyloseq object.
It can operate without any filters or apply a minimum sequence length filter before
calculating and plotting the distribution of sequence lengths.
}
\examples{
# Assuming 'ps' is a Phyloseq object
Go_SeqLengths(ps)
Go_SeqLengths(ps, filter = 100)

}
