% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_sumInfo.R
\name{Go_sumInfo}
\alias{Go_sumInfo}
\title{Summarize and Organize Text Information}
\usage{
Go_sumInfo(
  Extraction_Summary = NA,
  Taxa_overview = NA,
  Alpha_div = NA,
  Beta_div = NA,
  DA_test = NA,
  Conclusion = NA
)
}
\arguments{
\item{Extraction_Summary}{Path or vector of paths to the extraction summary text.}

\item{Taxa_overview}{Path or vector of paths to the taxa overview text.}

\item{Alpha_div}{Path or vector of paths to the alpha diversity summary text.}

\item{Beta_div}{Path or vector of paths to the beta diversity summary text.}

\item{DA_test}{Path or vector of paths to the differential abundance test summary text.}

\item{Conclusion}{Path or vector of paths to the conclusion text.}
}
\value{
A list containing the following elements:
\itemize{
\item{extraction_summary}{Path or vector of paths provided for the extraction summary text.}
\item{taxa_overview}{Path or vector of paths provided for the taxa overview text.}
\item{alpha_div}{Path or vector of paths provided for the alpha diversity summary text.}
\item{beta_div}{Path or vector of paths provided for the beta diversity summary text.}
\item{dA_test}{Path or vector of paths provided for the differential abundance test summary text.}
\item{conclusion}{Path or vector of paths provided for the conclusion text.}
}
}
\description{
This function organizes and returns the paths or vectors of paths for various summary texts related to nucleic acid extraction and sequencing.
If no paths are provided, it prints out a guide to using the function.
}
\details{
If no paths are provided, the function will print a guide explaining what each parameter expects and how to use the function.
This can be particularly helpful for users setting up the function for the first time.
}
\examples{
\dontrun{
sumInfo <- Go_sumInfo(
  Extraction_Summary = "text sentence",
  Taxa_overview = "text sentence",
  Alpha_div = "text sentence",
  Beta_div = "text sentence",
  DA_test = "text sentence",
  Conclusion = "text sentence"
)
print(summaryInfo)
}

}
