% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_tabByAntibiotic.R
\name{Go_tabByAntibiotic}
\alias{Go_tabByAntibiotic}
\title{Go_tabByAntibiotic}
\usage{
Go_tabByAntibiotic(df, antibiotics, SubjectID, save_CSV = FALSE)
}
\arguments{
\item{df}{A data frame containing the data to be analyzed.}

\item{antibiotics}{A character vector of antibiotic names corresponding to columns in \code{df}.}

\item{SubjectID}{A character string specifying the column name in \code{df} that contains the subject IDs.}

\item{save_CSV}{A logical value indicating whether to save the updated data frame as a CSV file. Default is FALSE.}
}
\value{
A data frame with additional columns indicating antibiotic usage by subject.
}
\description{
This function adds columns to a data frame indicating whether each subject used specific antibiotics. The function checks the usage of antibiotics by subject ID and adds a new column for each antibiotic with a "_byID" suffix. The updated data frame can also be saved as a CSV file.
}
\examples{
\dontrun{
df <- your_data_frame
antibiotics <- c("Antibiotic1", "Antibiotic2")
updated_df <- Go_tabByAntibiotic(df, antibiotics, SubjectID = "SubjectID")
}
}
