% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_tabByTimepoint.R
\name{Go_tabByTimepoint}
\alias{Go_tabByTimepoint}
\title{Go_tabByTimepoint}
\usage{
Go_tabByTimepoint(df, project, SubjectID, Timepoint, orders, filled_by = NULL)
}
\arguments{
\item{df}{A data frame containing the data to be analyzed.}

\item{SubjectID}{A character string specifying the column name in \code{df} that contains the subject IDs.}

\item{Timepoint}{A character string specifying the column name in \code{df} that contains the time points.}

\item{orders}{A character vector specifying the order of the time points. This determines the factor levels for the time points.}

\item{filled_by}{A character string specifying the column name in \code{df} that should be used to fill the time point values. If NULL, presence will be indicated by 0 and 1. Default is NULL.}
}
\value{
A data frame in wide format, summarizing the data by time points for each subject.
}
\description{
This function reshapes a data frame to summarize the presence or values at different time points for each subject. It can either mark presence with binary indicators (0 or 1) or use a specific variable's values to fill the time points.
}
\examples{
\dontrun{
df <- your_data_frame
orders <- c("Baseline", "Month1", "Month3")
df_wide <- Go_tabByTimepoint(df, SubjectID = "SubjectID", Timepoint = "Timepoint", orders = orders)
df_wide_with_values <- Go_tabByTimepoint(df, SubjectID = "SubjectID", Timepoint = "Timepoint", orders = orders, filled_by = "Measurement")
}

}
