% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_tabInfo.R
\name{Go_tabInfo}
\alias{Go_tabInfo}
\title{Retrieve and Organize Table Information}
\usage{
Go_tabInfo(
  ASVs_Tab = NA,
  Tract_Tab = NA,
  Alpha_divTab = NA,
  Alpha_div_LmerTab = NA,
  RNAseq = NA,
  Tab1 = NA,
  PermanovaTab = NA
)
}
\arguments{
\item{ASVs_Tab}{File path for the ASVs (Amplicon Sequence Variants) table in CSV format.}

\item{Tract_Tab}{File path for the sequencing QC (Quality Control) track table in CSV format.}

\item{Alpha_divTab}{Already loaded or computed alpha diversity table.}

\item{Alpha_div_LmerTab}{Already loaded or computed linear mixed-effects model results for alpha diversity.}

\item{PermanovaTab}{Already loaded or computed PERMANOVA (Permutational Multivariate Analysis of Variance) results.}
}
\value{
A list containing the following elements:
\itemize{
\item{asvs}{Data frame of ASVs loaded from \code{ASVs_Tab}, if available.}
\item{track}{Data frame of sequencing QC track loaded from \code{Tract_Tab}, if available.}
\item{adiv}{Alpha diversity table passed through \code{Alpha_divTab}.}
\item{lmer.tab}{Results of the alpha diversity linear mixed-effects model passed through \code{Alpha_div_LmerTab}.}
\item{Permanova}{PERMANOVA results passed through \code{PermanovaTab}.}
}
}
\description{
This function loads specified data tables from given file paths and returns a list of these tables.
If no file paths are provided, it outputs information about what each parameter expects.
}
\details{
If no file paths are provided, the function prints out what each parameter is intended for and returns nothing.
This is particularly useful for debugging or when setting up the function for the first time.
}
\examples{
\dontrun{
tabInfo <- Go_tabInfo(
  ASVs_Tab = "path/to/ASVs.csv",
  Tract_Tab = "path/to/tract.csv",
  Alpha_divTab = preloaded_alpha_diversity_table,
  Alpha_div_LmerTab = preloaded_lmer_results,
  PermanovaTab = preloaded_permanova_results
)
print(tabInfo)
}

}
