% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_tabTops.R
\name{Go_tabTops}
\alias{Go_tabTops}
\title{Convert CSV Data to Phyloseq Object}
\usage{
Go_tabTops(csv, project)
}
\arguments{
\item{csv}{The file path of the CSV file containing OTU data and taxonomy information.}

\item{project}{A string representing the name of the project.}
}
\value{
A phyloseq object created from the provided CSV data.
}
\description{
This function reads a CSV file containing OTU (Operational Taxonomic Unit) data and taxonomy information, then converts it into a phyloseq object.
}
\details{
The function reads the CSV file, splits the data into numeric (OTU) and non-numeric (taxonomy) matrices, and creates a phyloseq object. The phyloseq object is then saved as an RDS file in a directory named '2_rds'.
}
\examples{
# Example usage:
ps_object <- Go_tabTops(csv = "path/to/otu_data.csv", project = "MyMicrobiomeProject")

}
