% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_tabTops_markdown.R
\name{Go_tabTops_markdown}
\alias{Go_tabTops_markdown}
\title{Create Phyloseq Object from CSV File}
\usage{
Go_tabTops_markdown(csv)
}
\arguments{
\item{csv}{The path to the CSV file containing the microbiome data.
The CSV file should have a specific format: columns with numeric data representing OTUs and
columns with non-numeric data representing taxonomic information.}
}
\value{
A \code{phyloseq} object constructed from the OTU counts and taxonomic data in the CSV file.
}
\description{
This function reads a CSV file, separates the data into numeric and non-numeric components,
and constructs a \code{phyloseq} object. The numeric data is assumed to be OTU (Operational Taxonomic Unit)
counts, while the non-numeric data is treated as taxonomic information.
}
\examples{
# Assuming you have a CSV file 'my_data.csv' in the correct format:
ps_object <- Go_tabTops_markdown("path/to/my_data.csv")

}
