% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_volcanoPlot_v1.R
\name{Go_volcanoPlot}
\alias{Go_volcanoPlot}
\title{Generate Volcano Plots for Multiple Datasets}
\usage{
Go_volcanoPlot(
  project,
  file_path,
  files,
  fc,
  mycols = NULL,
  name,
  overlaps = 10,
  font,
  height,
  width
)
}
\arguments{
\item{project}{A character string representing the project name or identifier.}

\item{file_path}{The path to the directory containing the data files.}

\item{files}{A pattern to match the filenames for processing.}

\item{fc}{A numeric value representing the fold-change threshold for significance in the volcano plot.}

\item{mycols}{A character vector specifying custom colors for plotting.}

\item{name}{An optional character string to specify the name or identifier for the plot.}

\item{overlaps}{An integer value indicating the maximum number of overlaps for text labels in the plot.}

\item{font}{A numeric value specifying the font size for text elements in the plot.}

\item{height}{The height of the output PDF file for the plot.}

\item{width}{The width of the output PDF file for the plot.}
}
\value{
The function generates and saves volcano plot PDF files but does not return a value.
}
\description{
This function creates volcano plots for a series of datasets specified by file names.
It supports various data analysis tools (e.g., DESeq2, ALDEx2, ANCOMBC) and includes options
for color customization, labeling, and output file generation.
}
\examples{
# Assuming appropriate data files are in the specified directory
Go_volcanoPlot(project = "MyProject", file_path = "data/", files = "deseq2_results.csv",
               fc = 2, mycols = c("#00BFC4", "#F8766D"), name = "ExampleVolcanoPlot",
               overlaps = 10, font = 12, height = 7, width = 7)

}
