% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_volcanoPlot_markdown.R
\name{Go_volcanoPlot_markdown}
\alias{Go_volcanoPlot_markdown}
\title{Create Volcano Plots from Data Tables}
\usage{
Go_volcanoPlot_markdown(DAtab, fc, mycols = NULL, name, overlaps = 10, font)
}
\arguments{
\item{DAtab}{A list of data frames, each representing a dataset for which a volcano plot will be generated.}

\item{fc}{A numeric value representing the fold-change threshold for significance in the volcano plot.}

\item{mycols}{A character vector specifying custom colors for plotting.}

\item{name}{An optional character string to specify the name or identifier for the plot.}

\item{overlaps}{An integer value indicating the maximum number of overlaps for text labels in the plot.}

\item{font}{A numeric value specifying the font size for text elements in the plot.}
}
\value{
The function generates and prints volcano plots but does not return a value.
}
\description{
This function generates volcano plots for each data frame in a list of data frames.
It supports different types of data analysis tools (e.g., DESeq2, ALDEx2, ANCOMBC) and
provides options for color customization and text display.
}
\examples{
# Assuming 'data_list' is a list of data frames suitable for volcano plotting
Go_volcanoPlot_markdown(DAtab = data_list, fc = 2, mycols = c("#00BFC4", "#F8766D"), name = "ExamplePlot", overlaps = 10, font = 12)

}
