% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_zicoseq.R
\name{Go_zicoseq}
\alias{Go_zicoseq}
\title{Go_zicoseq}
\usage{
Go_zicoseq(
  psIN,
  cate.outs,
  group1,
  group2,
  rand.eff = NULL,
  con.vari = NULL,
  orders
)
}
\arguments{
\item{psIN}{A \code{phyloseq} object containing species count data along with sample and taxonomic metadata.}

\item{cate.outs}{A character string specifying the column in the sample metadata that defines the categorical groups for comparison.}

\item{group1}{The first group in the comparison.}

\item{group2}{The second group in the comparison.}

\item{rand.eff}{An optional character vector naming random effects variables to be used in the model, default is NULL.}

\item{con.vari}{An optional character string specifying a column in the sample metadata that should be used as a covariate in the model, default is NULL.}

\item{orders}{A character vector specifying the order of levels for the factor used in the model. This is particularly useful for setting the reference level in the analysis.}
}
\value{
A list containing the results of the ZicoSeq analysis, including p-values and effect sizes for features that are differentially abundant between the specified groups.
}
\description{
This function performs a differential abundance test using ZicoSeq on microbiome count data provided in a Phyloseq object. It filters data based on a categorical grouping and adjusts for covariates if provided.
}
\details{
The function subsets the \code{phyloseq} object to include only the samples of interest based on \code{cate.outs}. It then relevels the factor of interest to set the first specified group as the reference. It filters the OTU table to remove features with zero counts across all samples, and optionally adjusts for covariates and random effects if specified.
}
\examples{
\dontrun{
data(GlobalPatterns)
psIN <- GlobalPatterns
result <- Go_zicoseq(psIN, "SampleType", "Feces", "Soil", orders = c("Feces", "Soil"))
print(result)
}
}
