% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{GOEnrich_pannzer2}
\alias{GOEnrich_pannzer2}
\title{GO enrichment using pannzer2 result.}
\usage{
GOEnrich_pannzer2(
  deglist,
  pannzerfile,
  nodeSize = 1,
  useFDR = FALSE,
  cut = 0.05,
  outdir = NULL,
  outprefix = NULL
)
}
\arguments{
\item{deglist}{character vector of differentially expressed genes}

\item{pannzerfile}{GO.txt annotation file using pannzer2 database}

\item{nodeSize}{prune the GO hierarchy from the terms which have less than nodeSize annotated genes}

\item{useFDR}{whether to use FDR}

\item{cut}{significant pvalue threshold, if useFDR is set, use FDR instead}

\item{outdir}{output directory}

\item{outprefix}{output file prefix}
}
\description{
do GO enrichment by a list of DEGs using pannzer2 annotation output.
}
