\name{Internal functions}

\alias{getGoCategory}
\alias{getGOTerm}
\alias{getGOList}
\alias{goChildren}
\alias{parentsList}
\alias{isEndNode}
\alias{getGOID}
\alias{getGO.operon}
\alias{getGO.operon.main}
\alias{updateOligo2GO}
\alias{parentsListWraper}
\alias{parentsVectWraper}
\alias{ontoCompare.main}
\alias{goParents}
\alias{gowraper}

\title{Internal goTools functions}
 
\description{Internal goTools functions}
 
\details{These are not to be called by the user.}

\keyword{methods}
