% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsources.210324.R
\name{mcc}
\alias{mcc}
\title{Calculate Matthews Correlation Coefficient}
\usage{
mcc(preds = NULL, actuals = NULL, x = NULL, y = NULL)
}
\arguments{
\item{preds}{Predicted values (optional if \code{x} is provided).}

\item{actuals}{Actual values (optional if \code{y} is provided).}

\item{x}{Indicator matrix for predictions (if \code{preds} is not provided).}

\item{y}{Indicator matrix for truth (if \code{actuals} is not provided).}
}
\value{
Matthews Correlation Coefficient value.
}
\description{
Calculate Matthews Correlation Coefficient
}
\examples{
# Example usage pending specific data structure

}
