% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsources.210324.R
\name{multiplot}
\alias{multiplot}
\title{Plot Multiple ggplot Objects in a Single Layout}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, rows = 1)
}
\arguments{
\item{...}{ggplot objects to be plotted.}

\item{plotlist}{List of ggplot objects (alternative to ...).}

\item{file}{Filename if the plot is to be saved.}

\item{cols}{Number of columns in the plot layout.}

\item{rows}{Number of rows in the plot layout.}
}
\value{
A grid layout combining multiple ggplot objects.
}
\description{
Plot Multiple ggplot Objects in a Single Layout
}
\examples{
# Example usage with ggplot2 plots

}
