% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsources.210324.R
\name{normalizeByCols}
\alias{normalizeByCols}
\title{Normalize Data by Columns}
\usage{
normalizeByCols(df, csum = 1, level = NULL, delim = "\\\\|")
}
\arguments{
\item{df}{Data frame or matrix to be normalized.}

\item{csum}{Target sum for column normalization.}

\item{level}{Taxonomic level for normalization (for \code{normalizeByCols}).}

\item{delim}{Delimiter for splitting taxonomy strings (for \code{normalizeByCols}).}

\item{rsum}{Target sum for row normalization.}
}
\value{
Normalized data frame or matrix.
}
\description{
\code{normalizeByCols} normalizes the data by columns.
}
\examples{
data <- matrix(rnorm(20), nrow = 4)
normalizeByCols(data)

}
