% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsources.210324.R
\name{normalizeByRows}
\alias{normalizeByRows}
\title{Normalize Data by Rows or Columns}
\usage{
normalizeByRows(df, rsum = 1)
}
\arguments{
\item{df}{Data frame or matrix to be normalized.}

\item{rsum}{Target sum for row normalization.}

\item{csum}{Target sum for column normalization.}

\item{level}{Taxonomic level for normalization (for \code{normalizeByCols}).}

\item{delim}{Delimiter for splitting taxonomy strings (for \code{normalizeByCols}).}
}
\value{
Normalized data frame or matrix.
}
\description{
\code{normalizeByRows} normalizes the data by rows.
\code{normalizeByCols} normalizes the data by columns.
}
\examples{
data <- matrix(rnorm(20), nrow = 4)
normalizeByRows(data)
normalizeByCols(data)

}
