\name{probeID}
\alias{affylist}
\alias{operonlist}


\title{List of probe ids from Affymetrix hgu133a chip and Operon Version
  2 Human oligos}

\description{
  The \code{probeID} dataset consists of two lists of randomly chosen
  probe Ids. \code{affylist} contains 3 sets of ids from Affymetrix
  hgu133a. \code{operonlist} contains 2 sets of ids from Operon Version 2.}


\usage{data(probeID)}


\keyword{datasets}


