% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsources.210324.R
\name{removeRows}
\alias{removeRows}
\title{Remove specified rows from a data frame or matrix}
\usage{
removeRows(rowNum, data)
}
\arguments{
\item{rowNum}{A vector of row numbers to be removed.}

\item{data}{The data frame or matrix from which rows will be removed.}
}
\value{
A data frame or matrix with specified rows removed.
}
\description{
Remove specified rows from a data frame or matrix
}
\examples{
data <- data.frame(x = 1:5, y = 6:10)
removeRows(c(2, 4), data)

}
