% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsources.210324.R
\name{renameLevelsWithCounts}
\alias{renameLevelsWithCounts}
\title{Rename Levels with Counts}
\usage{
renameLevelsWithCounts(fvec, originalLevelsAsNames = FALSE)
}
\arguments{
\item{fvec}{Factor vector with levels to be renamed.}

\item{originalLevelsAsNames}{Logical; if TRUE, original levels are used as names.}
}
\value{
A factor vector with levels renamed to include counts.
}
\description{
Rename Levels with Counts
}
\examples{
vec <- factor(c("A", "B", "A", "C"))
renameLevelsWithCounts(vec)

}
