% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_Ancom2_v1.R
\name{Go_Ancom2}
\alias{Go_Ancom2}
\title{Perform ANCOM-II Analysis on Phyloseq Data}
\usage{
Go_Ancom2(
  psIN,
  project,
  data_type = "other",
  cate.outs,
  cate.conf = NULL,
  cont.conf = NULL,
  orders = NULL,
  name = NULL
)
}
\arguments{
\item{psIN}{Phyloseq object containing the data for analysis.}

\item{project}{Name of the project or analysis.}

\item{data_type}{Type of data in the Phyloseq object ("taxonomy", "kegg", "pathway", "RNAseq").}

\item{cate.outs}{Categorical outcomes to be analyzed.}

\item{cate.conf}{Categorical confounding variables to adjust for in the analysis.}

\item{cont.conf}{Continuous confounding variables to adjust for in the analysis.}

\item{orders}{Custom order of factors in the analysis, if applicable.}

\item{name}{Optional name for the analysis.}
}
\value{
The function generates CSV files containing the results of the ANCOM-II analysis.
These include statistics like log fold change, standard error, W statistic, p-values, and adjusted p-values,
along with taxonomic information for each feature. Files are saved in a specified directory
with a naming convention that includes key details of the analysis.
}
\description{
This function conducts Analysis of Compositions of Microbiomes (ANCOM-II) on data from a Phyloseq object.
It supports various types of variables including categorical and continuous, and allows for adjusting for confounders.
The function can handle multiple categories and generates output for each pairwise comparison.
}
\details{
The function first checks the type of data in the Phyloseq object.
It then processes categorical and continuous variables, adjusting for confounders if specified.
ANCOM-II analysis is performed for each variable or combination of variables.
The results are merged with taxonomy data and saved as CSV files.

For categorical variables, the function considers combinations of levels for analysis.
The analysis takes into account the structure of microbiome data and adjusts for compositionality.
}
\examples{
# psIN is a Phyloseq object
# Example usage:
Go_Ancom2(psIN = psIN,
          project = "MyProject",
          data_type = "taxonomy",
          cate.outs = c("Treatment", "Condition"),
          cate.conf = c("Gender"),
          cont.conf = NULL,
          orders = NULL,
          name = "Analysis1")

}
