% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_bdivMK_V1.R
\name{Go_bdivMK}
\alias{Go_bdivMK}
\title{Beta-diversity with MiRKAT / MiRKAT_LMM (facet annotation + CSV saving)}
\usage{
Go_bdivMK(
  psIN,
  cate.vars,
  project,
  orders,
  distance_metrics,
  cate.conf = NULL,
  plot = "PCoA",
  ellipse = TRUE,
  statistics = TRUE,
  mycols = NULL,
  rand.eff = NULL,
  combination = NULL,
  shapes = NULL,
  ID = NULL,
  facet = NULL,
  name = NULL,
  addnumber = TRUE,
  height,
  width
)
}
\arguments{
\item{psIN}{phyloseq object}

\item{cate.vars}{grouping variables to test (vector)}

\item{project}{project name (used in filenames)}

\item{orders}{factor order vector (optional)}

\item{distance_metrics}{e.g. c("bray","unifrac","wunifrac",...)}

\item{cate.conf}{covariates (character vector) \link{currently not used inside MiRKAT core test}}

\item{plot}{ordination method(s), e.g. "PCoA"}

\item{ellipse}{TRUE/FALSE or grouping name for ellipse color}

\item{statistics}{TRUE to compute MiRKAT/MiRKAT_LMM p-values}

\item{mycols}{manual color vector}

\item{rand.eff}{random-effect column for MiRKAT_LMM (e.g., "UID"); NULL → MiRKAT}

\item{combination}{NULL or 2/3/4 (subset pages; single PDF)}

\item{shapes}{shape mapping column (optional)}

\item{ID}{text label column (optional)}

\item{facet}{facet column (optional)}

\item{name}{extra name suffix (optional)}

\item{addnumber}{add n to group labels}

\item{height, width}{PDF size}
}
\value{
PDFs and CSVs written to disk
}
\description{
Beta-diversity with MiRKAT / MiRKAT_LMM (facet annotation + CSV saving)
}
