% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_biplot_function_V2.R
\name{Go_biplot}
\alias{Go_biplot}
\title{Generate NMDS Biplot for Microbiome Data}
\usage{
Go_biplot(
  psIN,
  metaData,
  project,
  orders,
  distance_metrics,
  data_type,
  biplot,
  shapes,
  TaxaLevel,
  ID,
  facet,
  name,
  height,
  width
)
}
\arguments{
\item{psIN}{Phyloseq object containing microbiome data.}

\item{metaData}{Path to the metadata CSV file.}

\item{project}{Name of the project for labeling.}

\item{orders}{Ordering of factors in the plot.}

\item{distance_metrics}{Vector of distance metrics to be used for NMDS.}

\item{data_type}{Type of data (e.g., 'dada2', 'Nephele', 'Other').}

\item{biplot}{Logical indicating whether to include biplot vectors.}

\item{shapes}{Shape of points in the plot.}

\item{TaxaLevel}{Taxonomic level for aggregation.}

\item{ID}{Identifier for labeling points.}

\item{facet}{Faceting variable for the plot.}

\item{name}{Optional name for the output plot.}

\item{height}{Height of the plot.}

\item{width}{Width of the plot.}
}
\value{
A PDF file containing the NMDS biplot.
}
\description{
Generate NMDS Biplot for Microbiome Data
}
\details{
This function creates NMDS (Non-metric Multidimensional Scaling) biplots for visualizing microbiome data, allowing for comparisons across various conditions or factors. The function supports different distance metrics, taxonomic levels, and customization options for the plot.
}
\examples{
Go_biplot(psIN = ps_object,
          metaData = "metadata.csv",
          project = "MyMicrobiomeStudy",
          orders = c("Condition1", "Condition2"),
          distance_metrics = c("bray", "unifrac"),
          data_type = "dada2",
          biplot = TRUE,
          shapes = 16,
          TaxaLevel = "Genus",
          ID = "SampleID",
          facet = "Group",
          name = "NMDSPlot",
          height = 6,
          width = 8)

}
