% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_blastASVs.R
\name{Go_blastASVs}
\alias{Go_blastASVs}
\title{Run BLAST on ASV Table Sequences}
\usage{
Go_blastASVs(project, asvsTable, blastDB)
}
\arguments{
\item{asvsTable}{The path to the CSV file containing ASV sequences with sequence IDs as row names.}

\item{blastDB}{The path to the BLAST database against which the sequences will be queried.
For example, "/Users/username/DB/blastDB/16S_ribosomal_RNA".}
}
\value{
Writes a CSV file to "1_out/updated_sequences_with_blast_results.csv"
with the original ASV data and added columns for BLAST hits.
The function also prints a message upon successful completion.
}
\description{
This function takes a CSV file containing ASV sequences, converts them to FASTA format,
runs BLAST searches against a specified database, parses the results, and returns
a CSV file with added BLAST hit information.
}
\details{
#===== download DB in local
cd DB
Download the nt db (300~500GB)
update_blastdb.pl nt
Download the RefSeq bacterial database
update_blastdb.pl --decompress refseq_rna
Or download the 16S microbial database
update_blastdb.pl --decompress 16S_ribosomal_RNA

The function processes each sequence individually, creates a temporary FASTA file,
runs a BLAST search, and parses the output to extract the hit number and hit definition.
It assumes that the BLAST tool is properly installed and accessible in the system's PATH.
The results include the genus and species extracted from the hit definitions.
}
\examples{
Go_blastASVs(asvsTable="path/to/your/asvTable.csv",
             blastDB="/path/to/your/blastDB/16S_ribosomal_RNA")

}
