% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_boxplot_V10.R
\name{Go_boxplot}
\alias{Go_boxplot}
\title{Generate Boxplots for Multiple Variables}
\usage{
Go_boxplot(
  df,
  cate.vars,
  project,
  outcomes,
  orders = NULL,
  mycols = NULL,
  combination = NULL,
  ylim = NULL,
  title = NULL,
  facet = NULL,
  paired = NULL,
  name = NULL,
  addnumber = TRUE,
  standardsize = TRUE,
  statistics = "yes",
  parametric = "no",
  star = "no",
  xangle = 90,
  cutoff = 0.1,
  height,
  width,
  plotCols,
  plotRows
)
}
\arguments{
\item{df}{Data frame containing the data to be plotted.}

\item{cate.vars}{Categorical variables to be used for the boxplot's x-axis.}

\item{project}{Project name used for output file naming.}

\item{outcomes}{Numeric variables to be plotted on the y-axis.}

\item{orders}{Vector of ordered factor levels for the categorical variables.}

\item{mycols}{Custom color palette for the plots.}

\item{combination}{Number of group combinations to display in the boxplot.}

\item{ylim}{Y-axis limits for the plot.}

\item{title}{Title of the plot.}

\item{facet}{Optional variable for creating facetted plots.}

\item{paired}{Indicates if the data points are paired.}

\item{name}{Optional name for saving the plot.}

\item{addnumber}{Boolean to add the number of samples in each group.}

\item{standardsize}{Boolean to control the size of the plot based on group size.}

\item{statistics}{Whether to perform statistical tests.}

\item{parametric}{Whether to use parametric tests.}

\item{star}{Whether to show significance levels as stars.}

\item{xangle}{Angle of x-axis labels.}

\item{cutoff}{Significance level for statistical tests.}

\item{height}{Height of the plot.}

\item{width}{Width of the plot.}

\item{plotCols}{Number of columns in the multiplot layout.}

\item{plotRows}{Number of rows in the multiplot layout.}
}
\value{
A PDF file containing the boxplots.
}
\description{
Generate Boxplots for Multiple Variables
}
\details{
This function creates boxplots for multiple categorical variables against one or more numeric outcomes. It supports faceting, custom color palettes, and statistical testing with options for parametric or non-parametric methods. The function can handle paired data and allows customization of plot size and layout.
}
\examples{
Go_boxplot(df = my_data, cate.vars = c("Group", "Condition"), project = "MyProject",
           outcomes = c("Variable1", "Variable2"), orders = c("Control", "Treatment"),
           height = 10, width = 8, plotCols = 2, plotRows = 1)

}
