% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_cluster_V2.R
\name{Go_cluster}
\alias{Go_cluster}
\title{Cluster ASVs to Form OTUs}
\usage{
Go_cluster(psIN, project, db, percent)
}
\arguments{
\item{psIN}{A phyloseq object containing ASV data.}

\item{project}{A string representing the name of the project.}

\item{db}{A string indicating the database file path for taxonomy assignment.}

\item{percent}{A numeric value representing the similarity percentage for clustering.}
}
\value{
A new phyloseq object with clustered OTUs and reclassified taxonomy.
}
\description{
This function clusters ASVs (Amplicon Sequence Variants) to form OTUs (Operational Taxonomic Units) based on a specified similarity percentage. It then reclassifies the clustered sequences and creates a new phyloseq object.
}
\details{
The function performs sequence alignment, calculates distance matrices, and identifies clusters of ASVs using the DECIPHER package. It then merges ASV counts within the same cluster and reassigns taxonomy using the specified database. The result is a new phyloseq object with updated OTU counts and taxonomy.
}
\examples{
# Example usage:
ps_clustered <- Go_cluster(psIN = ps, project = "MyProject", db = "path/to/db", percent = 97)

}
