% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_colbarchart_v3.R
\name{Go_colbarchart}
\alias{Go_colbarchart}
\title{Generate Color Bar Charts for Microbiome Data}
\usage{
Go_colbarchart(
  psIN,
  cate.vars,
  project,
  taxanames,
  data_type,
  orders,
  simple = FALSE,
  mycols = NULL,
  relative = T,
  x_label = NULL,
  facet = NULL,
  legend = "bottom",
  cutoff = 0.005,
  name = NULL,
  ncol = NULL,
  height,
  width
)
}
\arguments{
\item{psIN}{A phyloseq object containing microbiome data.}

\item{cate.vars}{A vector of categorical variables to be used for grouping the data in the bar charts.}

\item{project}{A string representing the project name, used for file naming.}

\item{taxanames}{A vector of taxonomic ranks to be visualized in the charts.}

\item{data_type}{A string indicating the type of data, affects taxonomy naming.}

\item{orders}{A vector specifying the order of categories within the plot facets.}

\item{simple}{A logical value, if TRUE, produces simpler plots without facets.}

\item{mycols}{Optional, custom color palette for the plot.}

\item{relative}{A logical value indicating whether to plot relative abundances.}

\item{x_label}{A string for the x-axis label, default is "SampleIDfactor".}

\item{facet}{A string or vector of variables to create facets in the plot.}

\item{legend}{Position of the legend, either "bottom" or "right".}

\item{cutoff}{A numeric value for filtering out low-abundance taxa.}

\item{name}{Optional, a string for additional naming in the plot title.}

\item{ncol}{Optional, number of columns in the facet layout.}

\item{height}{Plot height.}

\item{width}{Plot width.}
}
\value{
Creates and saves bar chart plots in PDF format in the specified output directory.
}
\description{
Produces color bar charts representing the relative or absolute abundance of microbial taxa, categorized by specified variables.
}
\details{
The function creates bar charts showing the abundance of microbial taxa, grouped by specified variables. It can handle both relative and absolute abundance data, and allows for flexible customization of plot appearance.
}
\examples{
# Example usage:
Go_colbarchart(psIN = my_phyloseq_object,
               cate.vars = c("Treatment", "TimePoint"),
               project = "MyMicrobiomeStudy",
               taxanames = c("Phylum", "Genus"),
               data_type = "16S",
               orders = c("Control", "Treatment"),
               relative = TRUE,
               height = 6,
               width = 8)

}
