% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_dualYplot_V1.R
\name{Go_dualYplot}
\alias{Go_dualYplot}
\title{Create a Dual Y-Axis Plot}
\usage{
Go_dualYplot(
  df,
  TaxTab,
  cate.vars,
  project,
  Box,
  Line1,
  Line2 = NULL,
  title = NULL,
  name = NULL,
  mycols = NULL,
  orders = NULL,
  xanlgle = 90,
  height,
  width
)
}
\arguments{
\item{df}{Data frame containing the data to be plotted.}

\item{TaxTab}{Path to the taxonomy table file.}

\item{cate.vars}{Vector of categorical variables for grouping.}

\item{project}{Name of the project associated with the data.}

\item{Box}{Column name in the data frame for boxplot values.}

\item{Line1}{Column name in the data frame for the first line plot values.}

\item{Line2}{Optional column name for the second line plot values.}

\item{title}{Optional title for the plot.}

\item{name}{Optional name for output file.}

\item{mycols}{Optional color palette for the plot.}

\item{orders}{Optional order of factor levels for the categorical variables.}

\item{height}{Height of the output plot.}

\item{width}{Width of the output plot.}

\item{xangle}{Rotation angle for x-axis labels.}
}
\value{
A PDF file containing the generated plot.
}
\description{
Create a Dual Y-Axis Plot
}
\details{
This function creates a dual y-axis plot, typically combining a boxplot and one or two line plots.
It is useful for visualizing relationships between different types of data (e.g., abundance and environmental variables) across categorical groups.
}
\examples{
Go_dualYplot(df = data_frame,
             TaxTab = "tax_table.csv",
             cate.vars = c("Group1", "Group2"),
             project = "MyProject",
             Box = "Abundance",
             Line1 = "Temperature",
             Line2 = "pH",
             title = "Abundance vs Environmental Factors",
             name = "Abundance_EnvFactors",
             height = 6,
             width = 8)

}
