% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_extendedBarplot.R
\name{Go_extendedBarplot}
\alias{Go_extendedBarplot}
\title{Extended Barplot from Phyloseq Data}
\usage{
Go_extendedBarplot(
  psIN,
  project,
  group1,
  group2,
  name = NULL,
  mvar,
  func,
  wilcox.p = 0.05,
  height,
  width
)
}
\arguments{
\item{psIN}{Phyloseq object containing microbiome data for analysis.}

\item{project}{Name of the project or analysis to label outputs.}

\item{group1}{First group for comparison, specified as a level within the metadata variable.}

\item{group2}{Second group for comparison, opposed to group1.}

\item{name}{Optional name for saving the plot, influences output file name.}

\item{mvar}{Metadata variable in Phyloseq object to divide data for comparison.}

\item{func}{Variable for dynamic grouping based on taxonomy or functional data ('KO', 'KO.des', 'pathway', 'path.des').}

\item{wilcox.p}{Cutoff for the p-value from the Wilcoxon test to determine statistical significance.}

\item{height}{Height of the output plot in inches.}

\item{width}{Width of the output plot in inches.}
}
\value{
Generates a PDF file containing the extended bar plot visualizing differences
between groups along with overlaid statistical significance. The function outputs
the modified data with significance levels and confidence intervals as well.
}
\description{
This function generates extended bar plots from a Phyloseq object for comparative analysis
between two groups at specified taxonomic or functional levels. It calculates statistical
significance using the Wilcoxon test and includes options for multiple transformations and
confidence interval calculations.
}
\details{
The function preprocesses the data to perform normalization transformations such as
Total Sum Scaling (TSS), relative abundance calculation, and Centered Log Ratio (CLR).
It then computes Wilcoxon tests for the specified groups across the data grouped by
'func' criteria. Results are visualized in an extended bar plot showing mean differences,
confidence intervals, and p-values for significant differences.
}
\examples{
# psIN is a Phyloseq object
# Example usage:
Go_extendedBarplot(psIN = psIN,
                   project = "MicrobiomeAnalysis",
                   group1 = "Control",
                   group2 = "Treatment",
                   name = "ExtendedBarplotExample",
                   mvar = "Group",
                   func = "pathway",
                   wilcox.p = 0.05,
                   height = 8,
                   width = 11)

}
