% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_function2ps_V2.R,
%   R/Go_function2ps_V2_humann3.R
\name{Go_function2ps}
\alias{Go_function2ps}
\title{Convert Functional Data to Phyloseq Object}
\usage{
Go_function2ps(tabPath, project = NULL, func.type, name = NULL)

Go_function2ps(tabPath, project = NULL, func.type, name = NULL)
}
\arguments{
\item{tabPath}{The path to the functional annotation table file.}

\item{project}{Optional project name for labeling purposes.}

\item{func.type}{The type of functional annotation, either PICRUSt or HUMAnN.}

\item{name}{Optional name for the output object.}
}
\value{
A \code{phyloseq} object containing the functional data. The function also saves an RDS file of the \code{phyloseq} object.

A Phyloseq object containing the functional annotation data, ready for ecological and bioinformatic analyses.
}
\description{
This function converts functional data from PICRUSt or HUMAnN into a \code{phyloseq} object for further analysis.
}
\details{
The function reads the functional data (e.g., KO or pathway abundances) and converts it into a format
suitable for integration into a \code{phyloseq} object. The function automatically detects the data type
(PICRUSt or HUMAnN) and processes the data accordingly.

This function converts a functional annotation table into a Phyloseq object for further analysis in ecological and bioinformatic contexts. It supports both PICRUSt and HUMAnN annotation types. The function reads the annotation table, cleans it by removing certain unwanted categories, and structures it into a format compatible with Phyloseq.
}
\examples{
Go_function2ps(tabPath = "path/to/functional_data.txt",
               project = "MyProject",
               func.type = "PICRUSt",
               name = "Analysis1")

ps_object <- Go_function2ps(tabPath = "path/to/annotation_table.csv",
                            project = "MyProject",
                            func.type = "PICRUSt",
                            name = "FunctionalAnalysis")

}
