% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_function2ps_V2.R
\name{Go_function2ps}
\alias{Go_function2ps}
\title{Convert Functional Data to Phyloseq Object}
\usage{
Go_function2ps(tabPath, project = NULL, func.type, name = NULL)
}
\arguments{
\item{tabPath}{The path to the functional data file.}

\item{project}{(Optional) A string indicating the project name for file naming.}

\item{func.type}{A string specifying the type of functional data ("PICRUSt" or "HUMAnN").}

\item{name}{(Optional) A string to add as a suffix to the file name.}
}
\value{
A \code{phyloseq} object containing the functional data. The function also saves an RDS file of the \code{phyloseq} object.
}
\description{
This function converts functional data from PICRUSt or HUMAnN into a \code{phyloseq} object for further analysis.
}
\details{
The function reads the functional data (e.g., KO or pathway abundances) and converts it into a format
suitable for integration into a \code{phyloseq} object. The function automatically detects the data type
(PICRUSt or HUMAnN) and processes the data accordingly.
}
\examples{
Go_function2ps(tabPath = "path/to/functional_data.txt",
               project = "MyProject",
               func.type = "PICRUSt",
               name = "Analysis1")

}
