% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_imgInfo.R
\name{Go_imgInfo}
\alias{Go_imgInfo}
\title{Retrieve and Organize Image Information}
\usage{
Go_imgInfo(
  Overview = NA,
  Rarefaction = NA,
  Barchart = NA,
  Bac.heatmap = NA,
  RNAseq.heatmap = NA,
  HumannHeatmap = NA,
  Adivplot = NA,
  Foreplot = NA,
  Bdivplot = NA,
  DAplot = NA,
  EBplot = NA,
  Network = NA,
  SHeatmap = NA,
  MAGs_tree = NA,
  MAGs_upset = NA,
  MAGs_heatmap = NA
)
}
\arguments{
\item{Rarefaction}{Path or vector of paths for rarefaction curve images.}

\item{Barchart}{Path or vector of paths for bar chart images.}

\item{Bac.heatmap}{Path or vector of paths for bacterial heatmap images.}

\item{Adivplot}{Path or vector of paths for alpha diversity plots.}

\item{Foreplot}{Path or vector of paths for forest plots.}

\item{Bdivplot}{Path or vector of paths for beta diversity plots.}

\item{DAplot}{Path or vector of paths for differential abundance plots.}
}
\value{
A list containing paths or vectors of paths for each type of graphical output.
}
\description{
This function organizes paths to various graphical outputs such as rarefaction curves,
bar charts, heatmaps, etc., and returns a list of these paths. The function checks if any
paths are provided and prints details about what each parameter expects if none are given.
}
\examples{
\dontrun{
  imgInfo <- Go_imgInfo(
    Rarefaction = "path/to/rarefaction.png",
    Barchart = c("path/to/bar1.png", "path/to/bar2.png"),
    Bac.heatmap = "path/to/heatmap.png"
  )
  print(imgInfo)
}

}
