% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_kraken2Extraction.V1.R
\name{Go_kraken2Extraction}
\alias{Go_kraken2Extraction}
\title{Go_kraken2Extraction}
\usage{
Go_kraken2Extraction(
  kraken2,
  bracken = NULL,
  kingdom = c("Bacteria", "Virus", "Fungi")
)
}
\arguments{
\item{kraken2}{Path to Kraken2 MPA file}

\item{bracken}{Optional Bracken file (required for Bacteria)}

\item{kingdom}{One of "Bacteria", "Virus", "Fungi"}
}
\value{
A phyloseq object
}
\description{
This function extracts taxonomy and abundance tables from Kraken2/Bracken
outputs and converts them into a standardized phyloseq object.
}
\details{
Parse Kraken2 (and optional Bracken) MPA output into a phyloseq object
with unified taxonomy for Bacteria, Virus, and Fungi.

Personal signature:
Heekuk Park, Columbia University Medical Center
Email: hp2523@cumc.columbia.edu
Date: 2025-11-01
}
\author{
Heekuk Park \href{mailto:hp2523@cumc.columbia.edu}{hp2523@cumc.columbia.edu}
}
