% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_mirkat_v1.R
\name{Go_mirkat}
\alias{Go_mirkat}
\title{Microbiome Regression-Based Kernel Association Test (MiRKAT)}
\usage{
Go_mirkat(psIN, project, cate.vars, cate.conf = NULL, orders, name = NULL)
}
\arguments{
\item{psIN}{Phyloseq object containing microbiome data.}

\item{project}{Name of the project, used for output file naming.}

\item{cate.vars}{List of categorical variables for MiRKAT analysis.}

\item{cate.conf}{Optional list of confounding variables.}

\item{orders}{A vector specifying the order of levels in the categorical variable.}

\item{name}{Optional name for the output file.}
}
\value{
The function generates a CSV file containing the MiRKAT analysis results, including P-values for the association between microbial communities and the categorical outcome. The file is saved in the specified output directory.
}
\description{
Microbiome Regression-Based Kernel Association Test (MiRKAT)
}
\details{
This function performs the MiRKAT (Microbiome Regression-Based Kernel Association Test) on microbiome data. MiRKAT is a non-parametric method used for assessing the association between microbial community composition and a categorical outcome variable, accounting for potential confounders.

The function handles multiple comparisons by considering different combinations of groups within the categorical variables and accounts for confounding factors if provided. It computes various UniFrac distances and converts them into kernel matrices for the MiRKAT analysis.
}
\examples{
Go_mirkat(psIN = my_phyloseq_object,
         project = "MyMicrobiomeProject",
         cate.vars = c("TreatmentGroup"),
         cate.conf = c("Age", "BMI"),
         orders = c("Control", "Treatment"),
         name = "MiRKAT_analysis")

}
