% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_mycols.R
\name{Go_myCols}
\alias{Go_myCols}
\title{Generate and Display Custom Color Palettes}
\usage{
Go_myCols(custumCols = NULL, RColorBrewer = NULL, piratepal = NULL)
}
\arguments{
\item{custumCols}{A character string specifying a predefined set of custom colors (e.g., 'cols1', 'cols2', 'cols3').}

\item{RColorBrewer}{A character string specifying a palette from the RColorBrewer package.}

\item{piratepal}{A character string specifying a palette from the piratepal function in the yarrr package.}
}
\value{
A vector of color values corresponding to the selected palette.
}
\description{
This function provides options for generating color palettes from predefined sets,
RColorBrewer palettes, or piratepal palettes. It displays a barplot of the selected
color palette and returns the color values.
}
\examples{
# Generate and display a palette from custom colors set 1
my_colors <- Go_myCols(custumCols = "cols1")

# Generate and display a palette from RColorBrewer
my_colors <- Go_myCols(RColorBrewer = "Set1")

# Generate and display a palette from piratepal
my_colors <- Go_myCols(piratepal = "basel")

}
