% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_network.R
\name{Go_network}
\alias{Go_network}
\title{Go_network}
\usage{
Go_network(
  tab1_path,
  tab2_path = NULL,
  tab3_path = NULL,
  Sampledata,
  mainGroup,
  subgroup,
  tab1_name,
  tab2_name = NULL,
  tab3_name = NULL,
  cutoff = 0.3,
  pval_threshold = NULL,
  qval_threshold = NULL,
  node_font,
  node_names = TRUE,
  name = NULL,
  width = 10,
  height = 10
)
}
\arguments{
\item{tab1_path}{Path to the CSV file containing the first table of data.}

\item{tab2_path}{Optional path to the CSV file containing the second table of data.}

\item{tab3_path}{Optional path to the CSV file containing the third table of data.}

\item{Sampledata}{Path to the CSV file containing sample metadata.}

\item{mainGroup}{The name of the main grouping variable in the sample metadata.}

\item{subgroup}{The name of the subgroup for filtering within the main group.}

\item{tab1_name}{Descriptive name for the first data table used in the network visualization.}

\item{tab2_name}{Optional descriptive name for the second data table.}

\item{tab3_name}{Optional descriptive name for the third data table.}

\item{cutoff}{Correlation coefficient threshold for including edges in the network.}

\item{pval_threshold}{Optional threshold for p-values to filter edges based on statistical significance.}

\item{qval_threshold}{Optional threshold for FDR-adjusted q-values to filter edges based on statistical significance.}

\item{node_font}{Font size for node labels in the network plot.}

\item{node_names}{Boolean to indicate whether to display node names or IDs.}

\item{name}{Optional name for additional naming in output files.}

\item{width}{Width of the output plot in inches.}

\item{height}{Height of the output plot in inches.}
}
\value{
Saves network plots as PDF files and optionally outputs centrality measures for nodes in the network.
}
\description{
Constructs and visualizes a network based on microbial community data across different samples, integrating multiple data tables.
}
\details{
This function reads microbial community data from specified CSV files, processes it, and constructs a network based on correlations among community members. It allows for extensive customization of the network visualization, including the ability to filter edges based on correlation coefficients, p-values, and q-values. Additionally, it can merge multiple tables and utilize sample metadata for grouping and subsetting the data.
}
\examples{
\dontrun{
Go_network(
  tab1_path = "path/to/data1.csv",
  tab2_path = "path/to/data2.csv",
  Sampledata = "path/to/sampledata.csv",
  mainGroup = "Treatment",
  subgroup = "Control",
  tab1_name = "Data Table 1",
  tab2_name = "Data Table 2",
  cutoff = 0.3,
  pval_threshold = 0.05,
  node_font = 12,
  node_names = TRUE,
  name = "NetworkPlot",
  width = 12,
  height = 12
)
}
}
