% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_pairedperm_V16.R
\name{Go_pairedperm}
\alias{Go_pairedperm}
\title{Perform Paired Permutational Multivariate Analysis of Variance (PERMANOVA)}
\usage{
Go_pairedperm(
  psIN,
  cate.vars,
  project,
  distance,
  distance_metrics,
  cate.conf = NULL,
  des,
  name = NULL
)
}
\arguments{
\item{psIN}{A \code{phyloseq} object containing the microbiome data.}

\item{cate.vars}{A vector of categorical variables to be tested.}

\item{project}{A string indicating the project name for directory and file naming.}

\item{distance}{A \code{phyloseq::distance} object or the method to calculate distances.}

\item{distance_metrics}{A vector of distance metrics to use for the PERMANOVA.}

\item{cate.conf}{(Optional) A vector of confounding variables to adjust for in the model.}

\item{des}{A description or label for the PERMANOVA run.}

\item{name}{(Optional) A name for the analysis for labeling and file naming.}
}
\value{
Outputs the results as CSV files in a specified directory. The function also returns the results as
a data frame.
}
\description{
This function conducts a paired PERMANOVA to assess differences in microbial community composition
between groups, taking into account paired samples or repeated measures.
}
\details{
The function performs a PERMANOVA analysis on each pair of levels within each categorical variable.
It can adjust for confounding variables and allows for multiple distance metrics.
}
\examples{
Go_pairedperm(psIN = ps,
              cate.vars = c("TreatmentGroup"),
              project = "MicrobiomeStudy",
              distance = "bray",
              distance_metrics = c("bray", "unifrac"),
              cate.conf = c("Age", "BMI"),
              des = "TreatmentComparison",
              name = "TreatmentAnalysis")

}
