% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gotool_dependency.R
\name{Gotool_dependency}
\alias{Gotool_dependency}
\title{Install and Load Necessary Packages for Gotools}
\usage{
Gotool_dependency()
}
\value{
This function does not return anything. It invisibly loads the required packages into the R session.
}
\description{
This function automates the process of installing and loading a predefined set of CRAN and Bioconductor packages necessary for the Gotools package. It checks if each package is already installed; if not, it installs the package and then loads it into the R session.
}
\details{
The function handles two types of packages: CRAN packages and Bioconductor packages. It first installs and loads CRAN packages, followed by Bioconductor packages. For Bioconductor package installation, \code{BiocManager} is used.

If \code{crayon} is installed, success messages are printed in color. The function is designed to be used at the beginning of a session to ensure all dependencies for Gotools are available.
}
\examples{
Gotool_dependency()

}
