\name{Internal functions}

\alias{ontoCompare.main}
\alias{getOntology}
\alias{getGOID}
\alias{isEndNode}
\alias{parentsVectWraper}
\alias{goParents}
\alias{goChildren}

\title{Internal goTools functions}
 
\description{Internal goTools functions}
 
\details{These are not to be called by the user.}

\keyword{methods}
